### 2024-08-17 #210 @pluflou
Solar Data Tools Submission
Submitting Author: Sara Miskovich (@pluflou)
All current maintainers: (@pluflou, @bmeyers)
Package Name: Solar Data Tools
One-Line Description of Package: Library of tools for analyzing photovoltaic power time-series data.
Repository Link:  https://github.com/slacgismo/solar-data-tools
Version submitted: 1.6.4
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15243716.svg)](https://doi.org/10.5281/zenodo.15243716)
Version accepted: 1.7.1

---

Solar Data Tools is an open-source Python library for analyzing PV power (and irradiance) time-series data. It provides methods for data I/O, cleaning, filtering, plotting, and analysis. These methods are largely automated and require little to no input from the user regardless of system type—from utility tracking systems to multi-pitch rooftop systems. Solar Data Tools was developed to enable analysis of unlabeled PV data, i.e. with no model, no meteorological data, and no performance index required, by taking a statistical signal processing approach in the algorithms used in the package’s main data processing pipeline.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

This package is for anyone dealing with photovoltaic data, especially data with no meteorological information (unlabeled). This includes photovoltaic professionals (in private solar industry or utility companies for example), researchers and students in the solar power domain, community solar owners, and anyone with a rooftop system. The scientific goal of the package is to facilitate analysis of photovoltaic data for any system, even those that are difficult to model, and the package uses signal decomposition to achieve that.

There are two other packages that are similar in that they offer data analysis tools for solar applications: [PVAnalytics](https://github.com/pvlib/pvanalytics) and [RdTools](https://github.com/NREL/rdtools). They are both model driven, and require the user to define their own analysis. PVAnalytics focuses on preprocessing and QA, while RdTools focuses on loss factor analysis. Solar Data Tools provides both data quality and loss factor analysis, runs _automatically_ with little to no setup, and is **model-free** and does not require any weather or other information. Solar Data Tools is most suited for when users want a pre-defined pipeline to get information on complex systems/sites that can't be modeled easily and that no meteorological data. A recent tutorial that was part of a virtual tutorial series on open-source tools and open-access solar data held by DOE’s Solar Technology Office in March 2024 goes over the differences in these packages and when each tool is appropriate to use. You can find the recording [here](https://www.youtube.com/watch?v=XKbqIlAEwOQ) and the slide deck [here](https://www.energy.gov/sites/default/files/2024-05/Data_Bounty_webinar_part_2.pdf) (see slide 16 for a summary).

https://github.com/pyOpenSci/software-submission/issues/204 (@cmarmo)

- [ ] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`. (will add soon)
