### 2024-08-19 d9914b6
starts paper for JOSS
