+++ .github/workflows/test-build.yml
@@ -69,6 +69,7 @@ jobs:
           echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
 
   deploy-rtd:
+    runs-on: ubuntu-latest
     steps:
   # Do not need to build and upload artifacts, just retrigger to update tags/version on website
       - name: Trigger RTDs build
