+++ .github/workflows/test-build.yml
@@ -70,6 +70,8 @@ jobs:
 
   deploy-rtd:
     runs-on: ubuntu-latest
+    needs: [build-pypi, build-conda]
+    environment: test-deployment
     steps:
   # Do not need to build and upload artifacts, just retrigger to update tags/version on website
       - name: Trigger RTDs build
