+++ .github/workflows/test-build.yml
@@ -70,10 +70,10 @@ jobs:
 
   deploy-rtd:
     runs-on: ubuntu-latest
-    needs: [build-pypi, build-conda]
+    #needs: [build-pypi, build-conda]
     environment: test-deployment
     steps:
-  # Do not need to build and upload artifacts, just retrigger to update tags/version on website
+    # Do not need to build and upload artifacts, just retrigger to update tags/version on website
       - name: Trigger RTDs build
         uses: dfm/rtds-action@v1
         with:
