+++ .github/workflows/test-build.yml
@@ -6,67 +6,67 @@ on:
   push:
     branches: [dev]
 jobs:
-  build-pypi:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v4
-
-      - uses: actions/setup-python@v4
-        with:
-          python-version: "3.10"
-
-      - name: Install Twine and Build
-        run: sudo pip install twine build
-
-      - name: Create the distribution
-        run: |
-          git fetch --prune --unshallow --tags
-          sudo python3 -m build
-
-  build-conda:
-    runs-on: ubuntu-latest
-    # sets default shell to remove need for source to run the conda shell
-    defaults:
-      run:
-        shell: bash -l {0}
-    environment: test-deployment
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v3
-
-      - uses: actions/setup-python@v4
-        with:
-          python-version: "3.10"
-
-      # Much better than manual installation, original version Miniconda2-4.7.10-Linux-x86_64.sh is broken
-      - uses: conda-incubator/setup-miniconda@v3
-        with:
-          auto-activate-base: true
-          activate-environment: ""
-          miniconda-version: "latest"
-
-      - name: Install the Conda Dependencies
-        run: |
-          conda config --set always_yes yes --set auto_update_conda false
-          conda update conda
-          conda install -n base conda-libmamba-solver
-          conda install python=3.10 conda-build colorama pip ruamel ruamel.yaml rich jsonschema -c conda-forge
-          git fetch --prune --unshallow --tags
-          pip install -e .
-
-      # echo yes before login to prevent anaconda bug breaking automation
-      # git tags MUST be fetched otherwise output will be blank
-      # bash variables cannot be used in github actions, must use actions specific syntax and methods
-      # channels need to be specified on build and are saved in the package for installs
-      - name: Build the Anaconda Package
-        id: condabuild
-        run: |
-          conda install anaconda-client
-          conda clean --all
-          conda config --set anaconda_upload no --set solver libmamba
-          VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)test conda build . -c mosek -c slacgismo -c conda-forge -c stanfordcvxgrp --numpy 2.0
-          echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
+#  build-pypi:
+#    runs-on: ubuntu-latest
+#    steps:
+#      - name: Checkout Repository
+#        uses: actions/checkout@v4
+#
+#      - uses: actions/setup-python@v4
+#        with:
+#          python-version: "3.10"
+#
+#      - name: Install Twine and Build
+#        run: sudo pip install twine build
+#
+#      - name: Create the distribution
+#        run: |
+#          git fetch --prune --unshallow --tags
+#          sudo python3 -m build
+#
+#  build-conda:
+#    runs-on: ubuntu-latest
+#    # sets default shell to remove need for source to run the conda shell
+#    defaults:
+#      run:
+#        shell: bash -l {0}
+#    environment: test-deployment
+#    steps:
+#      - name: Checkout Repository
+#        uses: actions/checkout@v3
+#
+#      - uses: actions/setup-python@v4
+#        with:
+#          python-version: "3.10"
+#
+#      # Much better than manual installation, original version Miniconda2-4.7.10-Linux-x86_64.sh is broken
+#      - uses: conda-incubator/setup-miniconda@v3
+#        with:
+#          auto-activate-base: true
+#          activate-environment: ""
+#          miniconda-version: "latest"
[…]
