### 2024-08-19 69fce4a
removes workflow trigger that was for testing
