+++ .github/workflows/test-build.yml
@@ -3,8 +3,6 @@ name: Test Build
 on:
   pull_request:
     branches: [main]
-  push:
-    branches: [dev]
 jobs:
   build-pypi:
     runs-on: ubuntu-latest
