### 2024-08-19 a5c646d
removes jobs, turns out I don't need these
