### 2024-08-19 b63c391
add signal decomposition to doc pages
