+++ paper/paper.md
@@ -47,29 +47,67 @@ output of the system.
 
 # Statement of need
 
+With the growing number
+of real-world installations of photovoltaic (PV) systems worldwide, 
+it is crucial to have tools that can process and analyze the data
+from system of all sizes and configurations. The data is 
+typically in the form of time-series measurements of real power 
+production reported as average power over some interval of time
+(between the scale of one minute to one hour) over a number of years, 
+possibly with missing entries. 
+
+
+
+Historically, PV data analysis tools have focused on data that are 
+combined with local meteorological measurements and system 
+configuration information (such as those from large power plants). The 
+data cleaning tasks have largely been manual, the analyses have 
+relied on metrics such as the performance index (townsend),
+which rely on having access to accurate site models and
+meteorological data.  In the case of smaller, 
+distributed rooftop PV systems, meteorological information is 
+typically lacking, and it's often difficult to model the system 
+accurately. For such systems, we desire to get insights from just 
+the PV power data in isolation (what we call \emph{unlabeled} data), 
+for which it is  difficult or impossible to form a performance 
+index. Given that distributed rooftop PV systems make up over 40\% of the  
+installed capacity in 2020 [@SEIA2021], there is a clear need for
+automated and model-free data processing and analysis tools that can 
+enable the remote monitoring of
+system health and optimization of operations and maintenance 
+activities of these systems.
+
+[//]: # (Cite dask IEEE short paper too? I don't mention cloud 
+deploymetn at all here)
+Solar Data Tools (SDT) [@Meyers2020b; @sdt-zenodo]
+is an open-source 
+Python library for automatic data processing and analysis of unlabeled 
+PV data signals. SDT automates the cleaning, filtering, and analyzing 
+PV power data, including loss factor estimation analysis, eliminating 
+the need for user configuration or 
+``babysitting" regardless of data quality or system configuration, 
+from large, utility-scale trackers to small, multi-pitch rooftops. 
+SDT provides practical tools for both small and 
+fleet-scale PV performance analyses, without the need
+to calculate performance indices for each system.
+
 There are two other libraries that are similar in that they offer data 
-analysis tools for solar applications: [PVAnalytics](https://github.
-com/pvlib/pvanalytics) and [RdTools](https://github.com/NREL/rdtools)
-. They are both model driven, and require the user to define their 
-own analysis. PVAnalytics focuses on preprocessing and QA, while 
-RdTools focuses on loss factor analysis. Solar Data Tools provides 
-both data quality and loss factor analysis, runs _automatically_ with little to no setup, and is **model-free** and does not require any weather or other information. Solar Data Tools is most suited for when users want a pre-defined pipeline to get information on complex systems/sites that can't be modeled easily and that no meteorological data. A recent tutorial that was part of a virtual tutorial series on open-source tools and open-access solar data held by DOE’s Solar Technology Office in March 2024 goes over the differences in these libraries and when each tool is appropriate to use. You can find the recording [here](https://www.youtube.com/watch?v=XKbqIlAEwOQ) and the slide deck [here](https://www.energy.gov/sites/default/files/2024-05/Data_Bounty_webinar_part_2.pdf) (see slide 16 for a summary).
-
-# Citations
-
-Citations to entries in paper.bib should be in
-[rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
-format.
-
-If you want to cite a software repository URL (e.g. something on GitHub without a preferred
-citation) then you can do it with the example BibTeX entry below for @fidgit.
-
-For a quick reference, the following citation commands can be used:
[…]
