+++ paper/paper.bib
@@ -1,0 +1,152 @@
+@article{pvanalytics,
+title = {PVAnalytics: A Python Package for Automated Processing of Solar Time Series Data},
+author = {Perry, Kirsten and Vining, William and Anderson, Kevin and Muller, Matthew and Hansen, Cliff},
+url = {https://www.osti.gov/biblio/1887283},
+place = {United States},
+year = {2022},
+month = {9}
+}
+
+@misc{signal-decomp-software,
+	author       = {Meyers, Bennet and others},
+	title        = {Signal decomposition modeling language},
+	month        = Sept,
+	year         = 2023,
+	version      = {v0.0.2},
+	url          = {https://github.com/cvxgrp/signal-decomposition}
+}
+
+@misc{rdtools-zenodo,
+	author       = {Deceglie, Michael and others},
+	title        = {NREL/rdtools: Version 3.0.0-alpha.6},
+	month        = August,
+	year         = 2024,
+	version      = {v3.0.0},
+	url          = {https://doi.org/10.5281/zenodo.13356337}
+}
+
+
+@misc{sdt-zenodo,
+  author       = {Meyers, Bennet and others},
+  title        = {slacgismo/solar-data-tools: v1.2.2},
+  month        = mar,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v1.2.2},
+  doi          = {10.5281/zenodo.10888385},
+  url          = {https://doi.org/10.5281/zenodo.10888385}
+}
+
+@misc{sdt-repo,
+	author       = {Meyers, Bennet and others},
+	title        = {Solar Data Tools},
+	month        = mar,
+	year         = 2024,
+	version      = {v1.2.2},
+	url          = {https://github.com/slacgismo/solar-data-tools}
+}
+
+@misc{pypi,
+  author = {Meyers, Bennet and others},
+  title = {Solar Data Tools on PyPi},
+  url = {https://pypi.org/project/solar-data-tools/},
+  year = {2024}
+}
+
+
+@misc{conda,
+  author = {Meyers, Bennet and others},
+  title = {Solar Data Tools on Anaconda},
+  url = {https://anaconda.org/slacgismo/solar-data-tools},
+  year = {2024}
+}
+
+
+@article{diamond2016cvxpy,
+  author  = {Steven Diamond and Stephen Boyd},
+  title   = {{CVXPY}: {A} {P}ython-embedded modeling language for convex optimization},
+  journal = {Journal of Machine Learning Research},
+  year    = {2016},
+  volume  = {17},
+  number  = {83},
+  pages   = {1--5},
+}
+
+@misc{mosek,
+   author = "MOSEK ApS",
+   title = "MOSEK Fusion API for Python 10.1.28",
+   year = 2024,
+   url = "https://docs.mosek.com/latest/pythonfusion/index.html"
+ }
+
+@article{Meyers2023,
+	year    = {2023},
+	volume  = {17},
+	journal = {Foundations and Trends in Signal Processing},
+	title   = {Signal Decomposition Using Masked Proximal Operators},
+	doi     = {10.1561/2000000122},
+	issn    = {1932-8346},
+	number  = {1},
+	pages   = {1-78},
+	author  = {B. Meyers and S. Boyd}
+}
+
+@article{volpatti2022,
+	title={Signal Decomposition via Quadratic-Separable Optimization},
+	author={Volpatti, Luke and Meyers, Bennet E and Boyd, Stephen P},
+	year={2022},
+	url={https://web.stanford.edu/~boyd/papers/sig_decomp_quad_sep.html}
+}
+
+@misc{Meyers2020b,
+	author={B. {Meyers} and E. {Apostolaki-Iosifidou} and L. {Schelhas}},
[…]
