### 2024-08-24 #210 @cmarmo
Hi @pluflou ! Thank you for submitting your package for pyOpenSci review.
Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

---
---

Solar Data Tools is in excellent condition! Congratulation for all your work! :rocket:

My only comment is related to the test coverage, which could be improved.
We don't set a minimal threshold for test coverage in order to start the review process, so, if you don't mind, just keep my comment somewhere as a reminder for future developments, and I'm going to look  right away for an editor.
