### 2024-08-24 #210 @pluflou
> I saw a new 1.6.2 version was released while waiting for my feedback: I took the liberty to update the version submitted so the reviewers would deal with the latest version available.

Thank you!!
