### 2024-08-31 #210 @cmarmo
Hi @pluflou , I am glad to announce that we have an editor for Solar Data Tools review.

@shirubana kindly accepted to take care of your submission. I am letting her introduce herself here and wishing a nice review process to all people involved.
