### 2024-09-04 7da82e1
adds workflow for publishing docker image
