+++ .github/workflows/publish-image.yaml
@@ -1,0 +1,27 @@
+name: Publish image to Docker Hub
+on:
+  [workflow_dispatch]
+#  workflow_run:
+#    workflows: [build]
+#    types:
+#      - completed
+
+jobs:
+  publish-image:
+    #if: ${{ github.event.workflow_run.conclusion == 'success' }}
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v3
+
+      - name: Login to Docker Hub
+        uses: docker/login-action@v2
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Build and Push Docker Image
+        run: |
+          docker build -t ./docker/. ${{ secrets.DOCKER_USERNAME }}/sdt-v2:latest
+          docker push ${{ secrets.DOCKER_USERNAME }}/sdt-v2:latest
