### 2024-09-04 9cc6805
adds workflow for publishing docker image
