### 2024-09-04 81d7c52
adds workflow for publishing docker image
