+++ .github/workflows/publish-image.yaml
@@ -4,7 +4,7 @@ on:
     branches:
       - deploy-docker
 #  workflow_run:
-#    workflows: [build]
+#    workflows: [ "Deploy Main"]
 #    types:
 #      - completed
 
@@ -25,5 +25,5 @@ jobs:
 
       - name: Build and Push Docker Image
         run: |
-          docker build -t ${{ secrets.DOCKER_USERNAME }}/sdt-v2:latest -f ./docker/Dockerfile .
-          docker push ${{ secrets.DOCKER_USERNAME }}/sdt-v2:latest
+          docker build -t ${{ secrets.DOCKER_USERNAME }}/sdt-v1:latest -f ./docker/Dockerfile .
+          docker push ${{ secrets.DOCKER_USERNAME }}/sdt-v1:latest
