### 2024-09-04 ac87a68
moves to slacgismo hub account
