### 2024-09-04 aa51c72
update docker docs and workflow
