+++ docker/sdt-v1_full_pip_freeze.txt
@@ -1,193 +1,0 @@
-solar-data-tools==1.6.0
-aiobotocore==2.13.1
-aiohttp==3.9.5
-aioitertools==0.11.0
-aiosignal==1.3.1
-anyio==4.4.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==2.4.1
-async-lru==2.0.4
-attrs==23.2.0
-azure-common==1.1.28
-azure-core==1.30.2
-azure-identity==1.17.1
-azure-mgmt-compute==31.0.0
-azure-mgmt-core==1.4.0
-azure-mgmt-network==25.4.0
-Babel==2.15.0
-beautifulsoup4==4.12.3
-bleach==6.1.0
-bokeh==3.4.2
-boto3==1.34.131
-botocore==1.34.131
-cachetools==5.4.0
-certifi==2024.7.4
-cffi==1.16.0
-charset-normalizer==3.3.2
-clarabel==0.9.0
-click==8.1.7
-cloudpickle==3.0.0
-comm==0.2.2
-contourpy==1.2.1
-cryptography==42.0.8
-cvxpy==1.5.2
-cycler==0.12.1
-dask==2024.5.2
-dask-cloudprovider==2022.10.0
-dask-expr==1.1.2
-debugpy==1.8.2
-decorator==5.1.1
-defusedxml==0.7.1
-distributed==2024.5.2
-ecos==2.0.14
-executing==2.0.1
-fastjsonschema==2.20.0
-fonttools==4.53.1
-fqdn==1.5.1
-frozenlist==1.4.1
-fsspec==2024.6.1
-google-api-core==2.19.1
-google-api-python-client==2.137.0
-google-auth==2.32.0
-google-auth-httplib2==0.2.0
-googleapis-common-protos==1.63.2
-graphviz==0.20.3
-h11==0.14.0
-h5py==3.11.0
-haversine==2.8.1
-hcloud==2.0.1
-httpcore==1.0.5
-httplib2==0.22.0
-httpx==0.27.0
-idna==3.7
-ipykernel==6.29.5
-ipython==8.26.0
-ipywidgets==8.1.3
-isodate==0.6.1
-isoduration==20.11.0
-jedi==0.19.1
-Jinja2==3.1.4
-jmespath==1.0.1
-joblib==1.4.2
-json5==0.9.25
-jsonpickle==3.2.2
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2023.12.1
-jupyter==1.0.0
-jupyter_client==8.6.2
-jupyter-console==6.6.3
-jupyter_core==5.7.2
-jupyter-events==0.10.0
-jupyter-lsp==2.2.5
-jupyter_server==2.14.2
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.4
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.11
-kiwisolver==1.4.5
-locket==1.0.0
-lxml==5.2.2
-lz4==4.3.3
-MarkupSafe==2.1.5
-matplotlib==3.9.1
-matplotlib-inline==0.1.7
-mistune==3.0.2
-msal==1.30.0
-msal-extensions==1.2.0
-msgpack==1.0.8
-multidict==6.0.5
-nbclient==0.10.0
-nbconvert==7.16.4
-nbformat==5.10.4
-nest-asyncio==1.6.0
-notebook==7.2.1
-notebook_shim==0.2.4
-numpy==2.0.0
-osqp==0.6.7.post0
[…]
