### 2024-09-15 #210 @shirubana
Ok, I think I am acquainted now with the steps/my job after perusing the guide and slack. I have started to look for reviewers.
