+++ paper/paper.md
@@ -27,83 +27,73 @@ bibliography: paper.bib
 
 Solar Data Tools is an open-source Python library for analyzing 
 photovoltaic (PV) power (and irradiance) time-series data. It 
-enables automated analysis of _unlabeled_ PV data, i.e. with no model, 
-no meteorological data, and no performance index required, by taking a 
-statistical signal processing approach in the algorithms used in 
-the library’s main data processing pipeline.
+enables automated analysis of _unlabeled_ PV data, meaning no 
+model, meteorological data, or performance index is required. The
+library uses signal decomposition techniques in its main data 
+processing pipeline.
 
-Solar Data Tools provides
+Solar Data Tools offers
 methods for data I/O, cleaning, filtering, plotting, and data 
-quality and loss analysis. These methods are largely automated and 
-require little to no input from the user regardless of system 
-type.  This library is for anyone dealing with photovoltaic data, 
-especially data with no meteorological information (unlabeled). This 
-includes photovoltaic professionals (in private solar industry or 
-utility companies for example), researchers and students in the 
-solar power domain, community solar owners, and anyone with a 
-rooftop system. The scientific goal of the library is to empower 
-PV system fleet owners or operators to analyze system performance, even 
-when they only have access to the most basic data stream—power 
-output of the system.
+quality and loss analysis. These methods are largely automated,
+requiring minimal user input regardless of the PV system 
+type.  This library is ideal for anyone working with PV data, 
+especially when meteorological information is unavailable. This 
+includes photovoltaic professionals in the private solar industry or 
+utility companies, researchers and students in the 
+solar power domain, community solar owners, and
+rooftop system owners. The primary goal of the library is to enable 
+PV system fleet owners or operators to analyze system 
+performance using only the basic power output data.
 
 # Statement of need
 
 With the growing number
 of real-world installations of photovoltaic (PV) systems worldwide, 
-it is crucial to have tools that can process and analyze the data
-from system of all sizes and configurations. The data is 
-typically in the form of time-series measurements of real power 
-production reported as average power over some interval of time
-(between the scale of one minute to one hour) over a number of years, 
-possibly with missing entries. 
+it is crucial to have tools that can process and analyze data
+from systems of all sizes and configurations. The data
+typically consists of time-series measurements of real 
+power production, reported as average power over intervals ranging from 
+one minute to one hour, spanning several years and possibly 
+containing missing entries.
 
-
-
-Historically, PV data analysis tools have focused on data that are 
-combined with local meteorological measurements and system 
-configuration information (such as those from large power plants). The 
-data cleaning tasks have largely been manual, the analyses have 
-relied on metrics such as the performance index (townsend),
-which rely on having access to accurate site models and
-meteorological data.  In the case of smaller, 
-distributed rooftop PV systems, meteorological information is 
-typically lacking, and it's often difficult to model the system 
-accurately. For such systems, we desire to get insights from just 
-the PV power data in isolation (what we call \emph{unlabeled} data), 
-for which it is  difficult or impossible to form a performance 
-index. Given that distributed rooftop PV systems make up over 40\% of the  
-installed capacity in 2020 [@SEIA2021], there is a clear need for
-automated and model-free data processing and analysis tools that can 
-enable the remote monitoring of
-system health and optimization of operations and maintenance 
-activities of these systems.
+Historically, PV data analysis tools have focused on data combined 
+with local meteorological measurements and system configuration 
+information, such as those from large power plants. Data cleaning tasks
+have largely been manual, and analyses have relied on metrics like the 
[…]
