+++ .github/workflows/test-build.yml
@@ -1,6 +1,7 @@
 name: Test Build
 
 on:
+  push:
   pull_request:
     branches: [main]
 jobs:
