+++ .github/workflows/test-build.yml
@@ -63,6 +63,7 @@ jobs:
         run: |
           conda install anaconda-client
           conda clean --all
-          conda config --set anaconda_upload no --set solver conda-libmamba-solver
+          conda update --all
+          conda config --set anaconda_upload no --set solver libmamba
           VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)test conda build . -c mosek -c slacgismo -c conda-forge -c stanfordcvxgrp --numpy 2.0
           echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
