+++ .github/workflows/test-build.yml
@@ -49,7 +49,6 @@ jobs:
         run: |
           conda config --set always_yes yes --set auto_update_conda false
           conda update conda
-          conda install -n base conda-libmamba-solver
           conda install python=3.10 conda-build colorama pip ruamel ruamel.yaml rich jsonschema -c conda-forge
           git fetch --prune --unshallow --tags
           pip install -e .
@@ -64,6 +63,6 @@ jobs:
           conda install anaconda-client
           conda clean --all
           conda update --all
-          conda config --set anaconda_upload no --set solver libmamba
+          conda config --set anaconda_upload no
           VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)test conda build . -c mosek -c slacgismo -c conda-forge -c stanfordcvxgrp --numpy 2.0
           echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
