+++ .github/workflows/test-build.yml
@@ -13,7 +13,7 @@ jobs:
 
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.12"
 
       - name: Install Twine and Build
         run: sudo pip install twine build
@@ -36,7 +36,7 @@ jobs:
 
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
+          python-version: "3.12"
 
       # Much better than manual installation, original version Miniconda2-4.7.10-Linux-x86_64.sh is broken
       - uses: conda-incubator/setup-miniconda@v3
@@ -49,7 +49,7 @@ jobs:
         run: |
           conda config --set always_yes yes --set auto_update_conda false
           conda update conda
-          conda install python=3.10 conda-build colorama pip ruamel ruamel.yaml rich jsonschema -c conda-forge
+          conda install python=3.12 conda-build colorama pip ruamel ruamel.yaml rich jsonschema -c conda-forge
           git fetch --prune --unshallow --tags
           pip install -e .
 
