+++ .github/workflows/build.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Install Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.12"
 
       - name: Install Twine and Build
         run: sudo pip install twine build
@@ -42,23 +42,20 @@ jobs:
 
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.12"
 
-      # Much better than manual installation, original version Miniconda2-4.7.10-Linux-x86_64.sh is broken
       - name: Install Miniconda
-        uses: conda-incubator/setup-miniconda@v2
+        uses: conda-incubator/setup-miniconda@v3
         with:
-          auto-activate-base: true
-          activate-environment: ""
+          activate-environment: true
+          python-version: 3.12
           miniconda-version: "latest"
+          channels: conda-forge,mosek,slacgismo
+          channel-priority: true
 
       - name: Install the Conda Dependencies
+        shell: bash -l {0}
         run: |
-          conda config --set always_yes yes --set auto_update_conda false
-          conda update conda
-          conda install -n base conda-libmamba-solver
-          conda install python=3.10 conda-build colorama pip ruamel ruamel.yaml rich jsonschema -c conda-forge
-          git fetch --prune --unshallow --tags
           pip install -e .
 
       # echo yes before login to prevent anaconda bug breaking automation
@@ -66,15 +63,15 @@ jobs:
       # bash variables cannot be used in github actions, must use actions specific syntax and methods
       # channels need to be specified on build and are saved in the package for installs
       - name: Build the Anaconda Package
+        shell: bash -l {0}
         id: condabuild
         run: |
-          conda install anaconda-client
-          conda config --set anaconda_upload no --set solver libmamba
-          echo yes | anaconda login --username ${{ secrets.ANACONDA_CLOUD_USERNAME }} --password ${{ secrets.ANACONDA_CLOUD_PASSWORD }}
-          VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-) conda build . -c mosek -c anaconda -c slacgismo -c conda-forge -c stanfordcvxgrp --numpy 2.0
+          conda install conda-build
+          VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)test conda build . --numpy 2.0
           echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
 
       - name: Upload the Anaconda Package
+        shell: bash -l {0}
         id: condaload
         run: |
           anaconda upload -u slacgismo  /usr/share/miniconda3/conda-bld/noarch/solar-data-tools-${{ steps.condabuild.outputs.gitversion }}-*.tar.bz2
