+++ .github/workflows/test-build.yml
@@ -68,4 +68,5 @@ jobs:
         shell: bash -l {0}
         id: condaload
         run: |
+          conda install anaconda-client
           anaconda upload -u smiskov  /usr/share/miniconda3/conda-bld/noarch/solar-data-tools-${{ steps.condabuild.outputs.gitversion }}-*.tar.bz2
