+++ .github/workflows/test-build.yml
@@ -69,4 +69,4 @@ jobs:
         id: condaload
         run: |
           conda install anaconda-client
-          anaconda upload -u slacgismo  /usr/share/miniconda3/conda-bld/noarch/solar-data-tools-${{ steps.condabuild.outputs.gitversion }}-*.tar.bz2
+          anaconda upload -u slacgismo  /home/runner/miniconda3/envs/true/conda-bld/noarch/solar-data-tools-test-py_0.tar.bz2
