+++ .github/workflows/test-build.yml
@@ -63,10 +63,3 @@ jobs:
           conda install conda-build
           VERSION_FROM_GIT_TAG=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)test conda build . --numpy 2.0 --no-anaconda-upload
           echo "gitversion=$(git tag --list "v*[0-9]" --sort=version:refname | tail -1 | cut -c 2-)" >> $GITHUB_OUTPUT
-
-      - name: Upload the Anaconda Package
-        shell: bash -l {0}
-        id: condaload
-        run: |
-          conda install anaconda-client
-          anaconda upload -u slacgismo  /home/runner/miniconda3/envs/true/conda-bld/noarch/solar-data-tools-test-py_0.tar.bz2
