+++ .github/workflows/test-build.yml
@@ -1,7 +1,6 @@
 name: Test Build
 
 on:
-  push:
   pull_request:
     branches: [main]
 jobs:
