+++ conda_recipe/meta.yaml
@@ -1,67 +1,0 @@
-{% set name = "solar-data-tools" %}
-
-package:
-  name: "{{ name|lower }}"
-  version: {{ environ.get('VERSION_FROM_GIT_TAG') }}
-
-source:
-  # git_url: https://github.com/slacgismo/solar-data-tools
-  path: .. # checked out repo instead of main repo branch
-
-build:
-  noarch: python
-  number: 0
-  script: {{ PYTHON }} -m pip install . --no-deps --ignore-installed -vv
-
-requirements:
-  host:
-    - pip
-    - python
-  run:
-    - pip
-    - python
-    - scipy
-    - numpy
-    - pandas
-    - scikit-learn
-    - jupyter
-    - matplotlib
-    - seaborn
-    - requests
-    - pvlib
-    - cvxpy
-    - pykml
-    - haversine
-    - smart_open
-    - sig-decomp
-    - clarabel
-    - qss
-    - tqdm
-    - spcqe
-    - dask
-    - distributed
-    - dask-cloudprovider
-    - graphviz
-    - bokeh
-
-test:
-  imports:
-    - solardatatools
-  commands:
-    - pip check
-  requires:
-    - pip
-
-about:
-  home: https://github.com/slacgismo/solar-data-tools
-  license: BSD-2-Clause
-  license_file: LICENSE
-  license_family: BSD
-  summary: Tools for performing common tasks on solar PV data signals
-  doc_url: https://solar-data-tools.readthedocs.io/
-
-extra:
-  recipe-maintainers:
-    - bmeyers
-    - thistleman
-    - pluflou
