+++ MANIFEST.in
@@ -1,8 +1,3 @@
-include solardatatools/fixtures/*
 prune statistical-clear-sky/fixtures
 prune notebooks
 prune docs
-prune tests
-prune data
-prune dataviewer
-prune conda_recipe
