### 2024-11-01 fb2b034
add __version__ attr
