### 2024-11-07 1e82025
update conda install docs
