+++ docs/source/index.rst
@@ -10,12 +10,12 @@
 
 Welcome to the Solar Data Tools Docs!
 =====================================
-|PyPI release| |Anaconda Cloud release|
+|PyPI release| |Conda-forge release|
 
 .. |PyPI release| image:: https://img.shields.io/pypi/v/solar-data-tools.svg
    :target: https://pypi.org/project/solar-data-tools/
-.. |Anaconda Cloud release| image:: https://anaconda.org/slacgismo/solar-data-tools/badges/version.svg
-   :target: https://anaconda.org/slacgismo/solar-data-tools
+.. |Conda-forge| image:: https://anaconda.org/conda-forge/solar-data-tools-feedstock/badges/version.svg
+   :target: https://anaconda.org/conda-forge/solar-data-tools-feedstock
 
 Solar Data Tools is an open-source Python library for analyzing PV power (and irradiance) time-series data.
 It was developed to enable analysis of unlabeled PV data, i.e. with no model, no meteorological data, and no performance
