### 2024-11-07 5e389d9
fix conda badges
