+++ docs/source/index.rst
@@ -7,14 +7,14 @@
     :width: 600
     :class: only-light
     :align: center
-
+https://github.com/enterprises/slacstanford
 Welcome to the Solar Data Tools Docs!
 =====================================
 |PyPI release| |Conda-forge release|
 
 .. |PyPI release| image:: https://img.shields.io/pypi/v/solar-data-tools.svg
    :target: https://pypi.org/project/solar-data-tools/
-.. |Conda-forge| image:: https://anaconda.org/conda-forge/solar-data-tools-feedstock/badges/version.svg
+.. |Conda-forge| image:: https://anaconda.org/conda-forge/solar-data-tools/badges/version.svg
    :target: https://anaconda.org/conda-forge/solar-data-tools-feedstock
 
 Solar Data Tools is an open-source Python library for analyzing PV power (and irradiance) time-series data.
