### 2024-11-07 e6d7074
fix conda badges
