+++ README.md
@@ -23,13 +23,10 @@
         <img src="https://img.shields.io/pypi/v/solar-data-tools.svg" alt="latest release" />
     </a>
     <a href="https://anaconda.org/conda-forge/solar-data-tools-feedstock">
-        <img src="https://anaconda.
-org/conda-forge/solar-data-tools/badges/version.svg" />
+        <img src="https://anaconda.org/conda-forge/solar-data-tools/badges/version.svg" />
     </a>
     <a href="https://anaconda.org/conda-forge/solar-data-tools-feedstock">
-        <img src="https://anaconda.
-org/conda-forge/solar-data-tools/badges/latest_release_date.
-svg" />
+        <img src="https://anaconda.org/conda-forge/solar-data-tools/badges/latest_release_date.svg" />
     </a>
 </tr>
 <tr>
@@ -74,12 +71,8 @@ com/slacgismo/solar-data-tools/blob/main/LICENSE">
 <tr>
     <td>Conda Downloads</td>
     <td>
-        <a href="https://anaconda.
-org/conda-forge/solar-data-tools-feedstock">
-            <img src="https://anaconda.
-org/conda-forge/solar-data-tools/badges/downloads.svg" 
-alt="conda-forge 
-downloads" />
+        <a href="https://anaconda.org/conda-forge/solar-data-tools-feedstock">
+            <img src="https://anaconda.org/conda-forge/solar-data-tools/badges/downloads.svg" alt="conda-forge downloads" />
         </a>
     </td>
 </tr>
