### 2024-11-07 66ca781
fix conda badges
