+++ docs/source/index.rst
@@ -15,7 +15,7 @@ Welcome to the Solar Data Tools Docs!
 .. |PyPI release| image:: https://img.shields.io/pypi/v/solar-data-tools.svg
    :target: https://pypi.org/project/solar-data-tools/
 .. |Conda-forge| image:: https://anaconda.org/conda-forge/solar-data-tools/badges/version.svg
-   :target: https://anaconda.org/conda-forge/solar-data-tools-feedstock
+   :target: https://anaconda.org/conda-forge/solar-data-tools
 
 Solar Data Tools is an open-source Python library for analyzing PV power (and irradiance) time-series data.
 It was developed to enable analysis of unlabeled PV data, i.e. with no model, no meteorological data, and no performance
