### 2024-11-13 #210 @pluflou
@cmarmo @shirubana is it possible to update to the latest version (1.6.4)? This version is now available on conda-forge (previous versions were on a private channel).
