### 2025-01-13 bb1529b
add authors and summary
