+++ paper/paper.md
@@ -11,65 +11,106 @@ authors: # TODO: generate full list of authors
     orcid: 0000-0002-3302-838X
     corresponding: true
     affiliation: 1
-  - name: Bennet Meyers
+  - name: Bennet E. Meyers
     orcid: 0000-0002-4089-0744
     corresponding: true
     affiliation: 1
+  - name: "Duncan Ragsdale"
+    affiliation: 1
+  - name: "Mitchell Victoriano"
+    affiliation: 1
+  - name: "Aramis Dufour"
+    affiliation: 2
+  - name: "Nimish Yadav"
+    affiliation: 3
+  - name: "Elpiniki Apostolaki-Iosifidou"
+    affiliation: 1
+  - name: "Claire Berschauer"
+    affiliation: 1
+  - name: "Chengcheng Ding"
+    affiliation: 3
+  - name: "Victor-Haoyang Lian"
+    affiliation: 3
+  - name: "Tristan Lin"
+    affiliation: 3
+  - name: "Junlin Luo"
+    affiliation: 3
+  - name: "Xiao Ming"
+    affiliation: 3
+  - name: "David Jose Florez Rodriguez"
+    affiliation: 3
+  - name: "Derin Serbetcioglu"
+    affiliation: 1
+  - name: "Jose St Louis"
+    affiliation: 3
+  - name: "Tadatoshi Takahashi"
+    affiliation: 1
+  - name: "Haoxi Zhang"
+    affiliation: 3
+  - name: "Jonathan Goncalves"
+    affiliation: 1
+  - name: "Nimish Telang"
+  - name: "Alejandro Londono-Hurtado"
+    affiliation: 1
+  - name: "Shixian Sheng"
+
 affiliations:
- - name: SLAC National Accelerator Laboratory, Menlo Park, CA, 94025, USA
-   index: 1
+  - name: SLAC National Accelerator Laboratory, Menlo Park, CA, 94025, USA
+    index: 1
+  - name: Stanford University, Stanford, CA, 94305, USA
+    index: 2
+  - name: Carnegie Mellon University, Pittsburgh, PA 15213, USA
+    index: 2
 date: 19 August 2024
 bibliography: paper.bib
 ---
 
 # Summary
+
 [//]: # (high level summary, for non expert audience)
 
-Solar Data Tools is an open-source Python library for analyzing 
-photovoltaic (PV) power (and irradiance) time-series data. It 
-enables automated analysis of _unlabeled_ PV data, meaning no 
-model, meteorological data, or performance index is required. The
-library uses signal decomposition techniques in its main data 
-processing pipeline.
+Effectively processing and leveraging the growing volume of photovoltaic (PV) system performance data is essential for
+the operation and maintenance of PV systems globally. However, many distributed rooftop PV systems suffer from lower
+data quality, are difficult to model, and lack access to reliable environmental data.
 
-Solar Data Tools offers
-methods for data I/O, cleaning, filtering, plotting, and data 
-quality and loss analysis. These methods are largely automated,
-requiring minimal user input regardless of the PV system 
-type.  This library is ideal for anyone working with PV data, 
-especially when meteorological information is unavailable. This 
-includes photovoltaic professionals in the private solar industry or 
-utility companies, researchers and students in the 
-solar power domain, community solar owners, and
-rooftop system owners. The primary goal of the library is to enable 
-PV system fleet owners or operators to analyze system 
-performance using only the basic power output data.
+Solar Data Tools is an open-source Python library designed for automated data quality and loss factor
+analysis of _unlabeled_ PV time-series data, i.e. without requiring a system model, meteorological 
+data, or performance indices. The primary objective of Solar Data Tools is to empower PV system operators or fleet 
+owners to understand their system's performance using only basic power output data.
+
[…]
