### 2025-01-13 4a69289
add pdf and alphabetize authors
