### 2025-01-17 455e59d
adding signal decomposition monograph
