+++ paper/paper.bib
@@ -100,3 +100,16 @@ year = {2021}
    volume = {1},
    year = {1994},
 }
+
+@article{Meyers2023,
+   author = {Bennet E. Meyers and Stephen P. Boyd},
+   doi = {10.1561/2000000122},
+   issn = {1932-8346},
+   issue = {1},
+   journal = {Foundations and Trends® in Signal Processing},
+   pages = {1-78},
+   title = {Signal Decomposition Using Masked Proximal Operators},
+   volume = {17},
+   url = {http://www.nowpublishers.com/article/Details/SIG-122},
+   year = {2023},
+}
