### 2025-01-17 4767aa4
edits and new references
