+++ paper/paper.bib
@@ -113,3 +113,73 @@ year = {2021}
    url = {http://www.nowpublishers.com/article/Details/SIG-122},
    year = {2023},
 }
+
+@inproceedings{Ogut2024,
+	author = {Giray Ogut and Bennet Meyers and Aramis Dufour and Stephen Boyd},
+	doi = {10.1109/PVSC57443.2024.10749393},
+	isbn = {978-1-6654-6426-0},
+	booktitle = {2024 IEEE 52nd Photovoltaic Specialist Conference (PVSC)},
+	month = {6},
+	pages = {877-883},
+	publisher = {IEEE},
+	title = {Time Dilated {B}undt Cake Analysis of {PV} Output},
+	url = {https://ieeexplore.ieee.org/document/10749393/},
+	year = {2024},
+}
+
+@INPROCEEDINGS{Pierce2024,
+  author={Pierce, Benjamin G. and Wieser, Raymond J. and Ciardi, Thomas G. and Yao, Ange Dominique and French, Roger H. and Bruckman, Laura S. and Li, Mengjie},
+  booktitle={2024 IEEE 52nd Photovoltaic Specialist Conference (PVSC)},
+  title={Comparison of Empirical and Data Driven Digital Twins for a PV+Battery Fleet},
+  year={2024},
+  volume={},
+  number={},
+  pages={1391-1397},
+  doi={10.1109/PVSC57443.2024.10749422}}
+
+@article{Lindig2021,
+author = {Lindig, Sascha and Moser, David and Curran, Alan J. and Rath, Kunal and Khalilnejad, Arash and French, Roger H. and Herz, Magnus and Müller, Björn and Makrides, George and Georghiou, George and Livera, Andreas and Richter, Mauricio and Ascencio-Vásquez, Julián and van Iseghem, Mike and Meftah, Mohammed and Jordan, Dirk and Deline, Chris and van Sark, Wilfried and Stein, Joshua S. and Theristis, Marios and Meyers, Bennet and Baumgartner, Franz and Luo, Wei},
+title = {International collaboration framework for the calculation of performance loss rates: Data quality, benchmarks, and trends (towards a uniform methodology)},
+journal = {Progress in Photovoltaics: Research and Applications},
+volume = {29},
+number = {6},
+pages = {573-602},
+keywords = {degradation rate, performance loss rate, PV system degradation, PV system performance},
+doi = {https://doi.org/10.1002/pip.3397},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pip.3397},
+year = {2021}
+}
+
+@conference{osti_1990039,
+  author       = {Perry, Kirsten and Meyers, Bennet and Muller, Matthew},
+  title        = {Survey of Time Shift Detection Algorithms for Measured PV Data},
+  annote       = {In this research, three variations of time shift detection algorithms were tested for their ability to detect time shift issues (including daylight savings time and random time shifts) in measured PV data sets. Two algorithms from the Python PVAnalytics package were assessed, and one algorithm from the Solar-Data-Tools package was assessed. Each algorithm's ability to accurately detect and measure time shifts was assessed.},
+  url          = {https://www.osti.gov/biblio/1990039},
+  place        = {United States},
+  organization = {National Renewable Energy Laboratory (NREL), Golden, CO (United States)},
+  year         = {2023},
+  month        = {07}}
+
+@article{Li2024,
+title = {Power modeling of degraded PV systems: Case studies using a dynamically updated physical model (PV-Pro)},
+journal = {Renewable Energy},
+volume = {236},
+pages = {121493},
+year = {2024},
+issn = {0960-1481},
[…]
