### 2025-01-18 466430e
compile joss paper and fix typos
