### 2025-01-21 ebc0a6d
solve dilation bug with nans
