### 2025-01-21 007e155
add option to use filled data matrix
