### 2025-01-21 #210 @pluflou
Hi @cmarmo @shirubana, happy new year! Just wanted to check in on how things are going, and have a small update. We have the JOSS paper in a branch's (`joss_paper`) root dir under `paper/` and I'm just wondering if we should merge that into main or if we should provide the link to the branch. Thank you!
