### 2025-02-04 #210 @cmarmo
@shirubana, don't worry! We all struggle with our schedules.

I'm fine with you as a reviewer as long as we can keep an open communication about the timeline.
Would you have time to finalize the review together with our first reviewer (or sooner... :) )?
Thank you for your help!
