### 2025-02-21 31d258b
adding the DEFAULT global back in because other functions referenced it.
