### 2025-03-11 #210 @cmarmo
I am happy to share that we have our second reviewer for Solar Data Tools! 🥳

@echedey-ls kindly accepted to help us with this review.

@echedey-ls, before beginning your review, please fill out our [pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us improve all aspects of our review and better understand our community. No personal data will be shared from this survey - it will only be used in an aggregated format by our Executive Director to improve our processes and programs.

Here is the [reviewers guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html). This guide contains all of the steps and information needed to complete your review.
Here is the [review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) that you will need to fill out and submit here as a comment, once your review is complete.

Please get in touch with any questions or concerns!

The deadline for this review for both reviewers is the end of April.

Thank you very much for your help @jinningwang and @echedey-ls!
