### 2025-03-13 #210 @echedey-ls
Hi everybody,
@cmarmo convinced me so here I am :D
I plan on opening issues in SDT where I find adequate, to reduce the amount of walls of texts I tend to make. I will be cross-linking and briefly describing them here. Lastly, I'll announce once the review template has been filled out and I consider my review finished. I can't tell whether I will fulfil the deadline time, but I'll try my best to do so.

<details><summary>Review - done 😄 </summary>

   - Some functions have empty docstrings, not a deal breaker but could be improved.
   - Not really, but not a blocker. The main example makes for a good starter of the repo.

- [x] Link to documentation website.

   - I don't think so, but the main functions in the example are well documented. Enough for an example-driven use of the project.

     I've not been able to do so, I don't have a MOSEK license.
      I wasn't able to do so, output below for version being reviewed; I guess it's due to being unable to set up AWS keys and MOSEK.

<details><summary>Details</summary>

```powershell
(venv) PS C:\Users\Yo\Documents\10_Programming\solar-data-tools> pytest -v --cov=src --cov-report=xml --cov-report=term-missing --color=yes tests/
==================================================== test session starts =====================================================
platform win32 -- Python 3.11.8, pytest-8.3.5, pluggy-1.5.0 -- C:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Scripts\python.exe
cachedir: .pytest_cache
rootdir: C:\Users\Yo\Documents\10_Programming\solar-data-tools
plugins: anyio-4.9.0, cov-6.1.1, github-actions-annotate-failures-0.3.0, github-report-0.0.1

tests/pvsystemprofiler/test_calculate_longitude.py::TestCalculateLongitude::test_calculate_longitude PASSED             [  2%]
tests/pvsystemprofiler/test_equation_of_time.py::TestEquationOfTime::test_eot_da_rosa PASSED                            [  4%]
tests/pvsystemprofiler/test_equation_of_time.py::TestEquationOfTime::test_eot_duffie PASSED                             [  7%]
tests/pvsystemprofiler/test_estimate_latitude.py::TestEstimateLatitude::test_estimate_latitude PASSED                   [  9%]
tests/pvsystemprofiler/test_fit_azimuth.py::TestFitAzimuth::test_fit_azimuth PASSED                                     [ 12%]
tests/pvsystemprofiler/test_fit_longitude.py::TestFitLongitude::test_fit_longitude PASSED                               [ 14%]
tests/pvsystemprofiler/test_fit_tilt.py::TestFitTilt::test_fit_tilt PASSED                                              [ 17%]
tests/solardatatools/test_clear_day_detection.py::TestClearDayDetection::test_clear_day_weights PASSED                  [ 19%]
tests/solardatatools/test_clear_day_detection.py::TestClearDayDetection::test_find_clear_days PASSED                    [ 21%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l1_l1d1_l2d2p365_default PASSED [ 24%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l1_l1d1_l2d2p365_idx_select PASSED [ 26%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d1_l2d2p365_default_long PASSED [ 29%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d1_l2d2p365_idx_select PASSED [ 31%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d1_l2d2p365_transition PASSED [ 34%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d1_l2d2p365_transition_wrong PASSED [ 36%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d1_l2d2p365_yearly_periodic PASSED [ 39%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_l2_l1d2_constrained_default PASSED [ 41%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_tl1_l2d2p365_default PASSED  [ 43%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cvx_tl1_l2d2p365_idx_select PASSED [ 46%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cxv_l2_l1d1_l2d2p365_default PASSED [ 48%]
tests/solardatatools/test_cvx_signal_decompositions.py::TestSignalDecompositions::test_cxv_tl1_l2d2p365_long_not_yearly_periodic PASSED [ 51%]
tests/solardatatools/test_data_filling.py::TestDataFilling::test_interp_missing PASSED                                  [ 53%]
tests/solardatatools/test_data_filling.py::TestDataFilling::test_zero_nighttime PASSED                                  [ 56%]
tests/solardatatools/test_data_handler.py::TestDataHandler::test_load_and_run FAILED                                    [ 58%]
tests/solardatatools/test_data_transforms.py::TestStandardizeTimeAxis::test_standardize_time_axis PASSED                [ 60%]
tests/solardatatools/test_data_transforms.py::TestMake2D::test_make_2d_no_freq PASSED                                   [ 63%]
tests/solardatatools/test_data_transforms.py::TestMake2D::test_make_2d_with_freq_set PASSED                             [ 65%]
tests/solardatatools/test_fix_time_shifts.py::TestFixTimeShift::test_fix_time_shifts PASSED                             [ 68%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l1_l1d1_l2d2p365_default PASSED      [ 70%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_default PASSED      [ 73%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_default_long PASSED [ 75%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_idx_select PASSED   [ 78%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_osqp PASSED         [ 80%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_yearly_periodic PASSED [ 82%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_yearly_periodic_sum_card PASSED [ 85%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d2_default PASSED               [ 87%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_tl1_l2d2p365_default PASSED          [ 90%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_tl1_l2d2p365_idx_select PASSED       [ 92%]
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_tl1_l2d2p365_long_not_yearly_periodic SKIPPED [ 95%]
tests/solardatatools/test_system_profiler.py::TestSystemProfiler::test_system_profiler FAILED                           [ 97%]
tests/solardatatools/test_utilities.py::TestOutlierFilter::test_basic_outlier_filter PASSED                             [100%] C
:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Lib\site-packages\coverage\inorout.py:509: CoverageWarning: Module src was never imported. (module-not-imported)
  self.warn(f"Module {pkg} was never imported.", slug="module-not-imported")
C:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Lib\site-packages\coverage\control.py:915: CoverageWarning: No data
was collected. (no-data-collected)
  self._warn("No data was collected.", slug="no-data-collected")

WARNING: Failed to generate report: No data to report.

C:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Lib\site-packages\pytest_cov\plugin.py:336: CovReportWarning: Failed to generate report: No data to report.

  warnings.warn(CovReportWarning(message), stacklevel=1)

========================================================== FAILURES ==========================================================
_____________________________________________ TestDataHandler.test_load_and_run ______________________________________________

self = <tests.solardatatools.test_data_handler.TestDataHandler testMethod=test_load_and_run>

    def test_load_and_run(self):
        filepath = Path(__file__).parent.parent
        data_file_path = filepath / "fixtures" / "data_transforms" / "timeseries.csv"
        df = pd.read_csv(data_file_path, parse_dates=[0], index_col=0)
        dh = DataHandler(df)
        dh.fix_dst()
        dh.run_pipeline(power_col="ac_power_01", fix_shifts=True, verbose=False)
        # dh.report()
        self.assertAlmostEqual(dh.capacity_estimate, 6.7453649044036865, places=2)
        self.assertAlmostEqual(dh.data_quality_score, 0.9948186528497409, places=3)
        self.assertAlmostEqual(dh.data_clearness_score, 0.49222797927461137, places=3)
>       self.assertTrue(dh.inverter_clipping)
E       AssertionError: None is not true

C:\Users\Yo\Documents\10_Programming\solar-data-tools\tests\solardatatools\test_data_handler.py:21: AssertionError
__________________________________________ TestSystemProfiler.test_system_profiler ___________________________________________

self = <tests.solardatatools.test_system_profiler.TestSystemProfiler testMethod=test_system_profiler>

    def test_system_profiler(self):
        filepath = Path(__file__).parent.parent
        data_file_path = (
            filepath / "fixtures" / "system_profiler" / "data_handler_input.csv"
        )
        data = pd.read_csv(data_file_path, parse_dates=[1], index_col=1)
        dh = DataHandler(data)
        dh.fix_dst()
        dh.run_pipeline(
            power_col="ac_power", fix_shifts=False, correct_tz=False, verbose=False
        )
        # dh.report()
>       dh.setup_location_and_orientation_estimation(-5)

C:\Users\Yo\Documents\10_Programming\solar-data-tools\tests\solardatatools\test_system_profiler.py:21:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
C:\Users\Yo\Documents\10_Programming\solar-data-tools\solardatatools\data_handler.py:1501: in setup_location_and_orientation_estimation
    est = ConfigurationEstimator(
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <pvsystemprofiler.estimator.ConfigurationEstimator object at 0x000002A94C9C67D0>
data_handler = <solardatatools.data_handler.DataHandler object at 0x000002A94CC7B6D0>, gmt_offset = -5
solar_noon_method = 'optimized_estimates', daylight_method = 'optimized_estimates', data_matrix = 'filled'
daytime_threshold = 0.001

    def __init__(
        self,
        data_handler,
        gmt_offset,
        solar_noon_method="optimized_estimates",
        daylight_method="optimized_estimates",
        data_matrix="filled",
        daytime_threshold=None,
    ):
        if not data_handler._ran_pipeline:
            data_handler.run_pipeline()
        self.data_handler = data_handler
        if data_matrix == "raw":
            self.data_matrix = data_handler.raw_data_matrix
        elif data_matrix == "filled":
            self.data_matrix = data_handler.filled_data_matrix
        # Parameters to be estimated
        self.longitude = None
        self.latitude = None
        self.tilt = None
        self.azimuth = None
        # Attributes used for all calculations
        self.gmt_offset = gmt_offset
        self.hours_daylight = None
        self.daily_meas = self.data_handler.filled_data_matrix.shape[0]
        self.daytime_threshold = None
        self.day_interval = None
        self.x1 = None
        self.x2 = None
        self.data_sampling = self.data_handler.data_sampling
        self.num_days = self.data_handler.num_days
        self.day_of_year = self.data_handler.day_index.dayofyear
        self.eot_duffie = eot_duffie(self.day_of_year)
        self.eot_da_rosa = eot_da_rosa(self.day_of_year)
        self.delta = delta_cooper(self.day_of_year, self.daily_meas)
        self.omega = None
        self.days = np.logical_and(
>           data_handler.daily_flags.clear, ~data_handler.daily_flags.inverter_clipped
        )
E       TypeError: bad operand type for unary ~: 'NoneType'

C:\Users\Yo\Documents\10_Programming\solar-data-tools\pvsystemprofiler\estimator.py:77: TypeError
====================================================== warnings summary ======================================================
tests/solardatatools/test_data_filling.py::TestDataFilling::test_zero_nighttime
tests/solardatatools/test_data_filling.py::TestDataFilling::test_zero_nighttime
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_osqp
  C:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Lib\site-packages\osqp\interface.py:290: DeprecationWarning: "polish" is deprecated. Please use "polishing" instead.
    warnings.warn(

tests/solardatatools/test_data_filling.py::TestDataFilling::test_zero_nighttime
tests/solardatatools/test_data_filling.py::TestDataFilling::test_zero_nighttime
tests/solardatatools/test_signal_decompositions.py::TestSignalDecompositions::test_l2_l1d1_l2d2p365_osqp
  C:\Users\Yo\Documents\10_Programming\solar-data-tools\venv\Lib\site-packages\osqp\interface.py:405: PendingDeprecationWarning: The default value of raise_error will change to True in the future.
    warnings.warn(

======================================================= tests coverage =======================================================
______________________________________ coverage: platform win32, python 3.11.8-final-0 _______________________________________

================================================== short test summary info ===================================================
FAILED tests/solardatatools/test_data_handler.py::TestDataHandler::test_load_and_run - AssertionError: None is not true
FAILED tests/solardatatools/test_system_profiler.py::TestSystemProfiler::test_system_profiler - TypeError: bad operand type for unary ~: 'NoneType'
=============================== 2 failed, 38 passed, 1 skipped, 6 warnings in 74.48s (0:01:14) ===============================
```

     - Just in case, I can't state whether or not the range of inputs and conditions is enough.
         Package lacks linting CI, but this is not a deal breaker. May be yet another improvement in the maintenance side of it. Style seems consistent enough.

Estimated hours spent reviewing: 18-24

---

Some points need to be made regarding the project.
The first one is that it fulfils one of the most important needs in the solar industry. Both congrats and thanks to the authors.

There are obviously things to improve, my main concern is on long-term maintenance and how it the project will age. A second concern is related to devops, but since that impacts the user, I'm sure they'll get fixed.

I've opened a bunch of issues, cross-referencing here for information and commodity:
- https://github.com/slacgismo/solar-data-tools/issues/183
- https://github.com/slacgismo/solar-data-tools/issues/184
- https://github.com/slacgismo/solar-data-tools/issues/185
- https://github.com/slacgismo/solar-data-tools/issues/186
- https://github.com/slacgismo/solar-data-tools/issues/191
- https://github.com/slacgismo/solar-data-tools/issues/194
- https://github.com/slacgismo/solar-data-tools/issues/195

I'll leave some more other ideas of improvable thingies down below, not a requirement to pass my review:
- Add an optional `ax` parameter in `plot_*` functions of classes (DataHandler is my main concern, haven't dug into the others) that is passed directly to `solardatatools.plotting.plot_2d`. It would ease the creation of a complex data analysis for many sites, for example, by easing the generation of multi-axes plots, or web-based services.
- For ease of use, I would put all packages distributed under one same namespace (so they would instead become modules of a common package), for convention with most Python packages. It can be à la scipy, in which they only allow importing submodules of scipy, so a plain scipy import does not import everything and is faster (I guess the distinction in SDT was done for performance reasons?). This is clearly a breaking change.
- Again, convention blah blah blah, `src/` repo layout is probably beneficial.
- There shouldn't be the need to duplicate anything, I'm looking at `notebooks\examples\DataHandler_Demo-Default.ipynb` vs `docs\source\getting_started\notebooks\demo_default.ipynb` (the latter has at least one plot at the end the former does not). You can reference the notebooks at `/notebooks` from docs with a relative path.

Circling back to the long-term maintenance concern, I miss more comments explaining what the hidden parts of the code do, and remove commented-out code blocks if they are unnecessary. Yet again (I think I've written it in some issue), I do heavily recommend an extensive page with devops like SDT version bumping, python version addition/removal, deprecation procedures (if needed), Docker troubleshooting and/or Dask specific instructions. Thinking aloud.

Just verbosing ideas. The final point is a big congrats on the project. Super insightful and probably a game changer for distributed systems maintenance.
