### 2025-03-13 #210 @cdeline
Hi.  I look forward to following @echedey-ls 's example as I learn this process. I will be submitting review pull requests from my fork at https://github.com/cdeline/solar-data-tools.  My reviewer template is copied below:

  <summary>Review - will be updated over time </summary>

~- [ ] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).~
I do have to disclose that I've worked with Bennet in the past, have coauthored papers, and will be working more closely with him in the future.  I hope that this doesn't affect my objectivity in my review.

- [ ] **Performance:** Any performance claims of the software been confirmed. **(N/A)**
  **(some unit tests are included - overall test coverage of 35% isn't great, but I won't hold up a review just because of low test coverage.  It seems like there could be some easy tests that could be added just to confirm functionality)**

    - [x] Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)  **(yes - code analysis in Spyder returns code formatting scores 6/10 or better)**

The package contains a `paper.md`  **(yes - in the `joss_paper` branch)** matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:

Estimated hours spent reviewing: 5-6

---
