### 2025-03-13 slacgismo/solar-data-tools#183 @echedey-ls
[pyOpenSci reviewer] Unmaintained CHANGELOG.md file
**Issue**
`CHANGELOG.md` does not seem to be maintained.

File: https://github.com/slacgismo/solar-data-tools/blob/main/CHANGELOG.md

Last updated: 2019-05-15 11:13:58 -0700

**Suggestion**
This objectively does not impact the usability of the repo, but for maintenance purposes, I think it should be removed.
