### 2025-03-13 slacgismo/solar-data-tools#184 @echedey-ls
[pyOpenSci reviewer] Inconsistency between Python supported versions
**Issue**
Python version requirement and metadata in `pyproject.toml` do not match.

Requirement: `requires-python = ">=3.9"`

However, metadata classifiers state from 3.6 to 3.8:

```
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
```

Confirmed for latest version.

**Suggestion 1**
Update the classifiers.

**Suggestion 2**
Create a page with maintenance instructions, and in a section about bumping minimal Python requirement, add a step to search for the python version string being removed. For example, executing:
```
git grep -e '3\.xx' -- ':*.rst' ':*.md' ':*.toml' ':*.yaml'
```
