### 2025-03-13 slacgismo/solar-data-tools#185 @echedey-ls
[pyOpenSci reviewer] Tests missing Python version(s) in the matrix
**Issue**
https://github.com/slacgismo/solar-data-tools/blob/813b5054b879edb8468f70e2aada0628fed3e5b0/.github/workflows/test.yml#L18C1-L19C49

While `pyproject.toml` supports Python >=3.9, the package is tested against only 3.10, 3.11, 3.12 (this is true for the latest version too).
I've installed all the dependencies in Python 3.9 for SDT version 1.6.4 and did not found any problems (I still have to run the tests, thou).

**Suggestion 1**
Add 3.9 to the test matrix.

**Suggestion 2**
Add 3.13 to the test matrix (plus to the trove classifiers in `pyproject.toml`).

I don't know if this was deliberate, but in case it was not, to avoid problems in the future.
