### 2025-03-13 slacgismo/solar-data-tools#186 @echedey-ls
[pyOpenSci reviewer] De-duplication of Docker requirements and other Docker improvements
I understand the main purpose of the docker image is to run a dask API of SDT on servers and improve the computation time (disclaimer: I haven't dug deep to try/have a look at it yet, nor used dask in my life).

I wonder how the described issue may impact maintenance for the future, but in any case, this is not a dealbreaker for me as a reviewer. I suspect the user base that uses dask is marginal.

**Issue**
There's a triplication of requirements for the `[dask]` optional project dependencies:
1. The first one is in the `pyproject.toml` (I agree this one should remain)
2. The second one is in `docker/requirements.txt`, with the following changes:
   1. It specifies `solar-data-tools` as an external dependency, so it uses the latest SDT published in PyPI and not the local one (dockerizing SDT from a local commit is not straightforward, see it was described in `docker/README.md` for a remote branch)
      - I describe alternatives down below.
   3. A new dependency is added, `boto3==1.34.131`
   4. Versions are specified for `graphviz==0.20.3` and `bokeh==3.4.2`
5. Lastly, in `docker/README.md`. I think it would be better to tell users to install via `pip install solar-data-tools[dask]`, instead of detailing all the dependencies and the versions.

**Mix of suggestions**
#### Regarding de-duplication
1. Remove requirements.txt under docker/
2. Modify Dockerfile to:
   - If doing a local installation:
      - Copy the repo content
      - Run `pip install .[dask]`
      - Remove the copied repo folder, as the package has already been installed in the image (I mean, it's not an editable install)
   - If users are not expected to have cloned the repo, provide `pip install solar-data-tools[dask]`
   - Run `pip install boto3==1.34.131`

#### Regarding the provided Docker instructions
I feel like this is just pure convenience for the developers using it, and that's okay. However, if you already are providing this image to the users, I prefer to see a How To Use the docker image as a whole section (in the sphinx docs) and another section maybe, on How to build your own docker image, than a mixture of both and development tactics.

In addition to #184, I miss a section with concrete steps for debugging, releasing; that could hold parts of docker/README.md

And here ends my first wall of text 😬
