### 2025-03-16 slacgismo/solar-data-tools#191 @echedey-ls
[pyOpenSci reviewer] PyPI description broken image - MANIFEST.in
**Issue**
Image in PyPI description (same as README.md file) is not shown. See https://pypi.org/project/solar-data-tools/ .

This happens because the docs directory is not included in the source distribution.

**Solution**
`MANIFEST.in` should include file `docs/source/_static/SDT_v1_secondary_blue_text.png`

**Recommendation**
A source distribution is expected to have all the files necessary for another developer to use the package, and build it themselves. By default, all users will install the wheel, which only includes necessary files of the package, so inclusion of extra files is not an expected problem. Tests are recommended so, by default, users make use of the minimal amount of files required, but if they want, they can test the package by using the source distribution. For documentation, it's almost the same. For practicality, the needed files to build the documentation could be included. The latter solves this issue.

My recommendation is to remove the following two lines in `MANIFEST.in`:

```
prune docs
prune tests
```

**Context addition**
Minor tweaks to the same `MANIFEST.in` that I won't open an issue for it:
- the path `statistical-clear-sky/fixtures` is outdated as it no longer exists
- `docker/` can be optionally pruned
- `.github/`, `.pre-commit-config.yaml`, `.readthedocs.yaml`, `CODE_OF_CONDUCT.md`, are usually pruned
- `requirements.txt` is not even necessary since it is redundant with `pyproject.toml`
