### 2025-03-16 slacgismo/solar-data-tools#184 @echedey-ls
To add, typing semantics that all the project widely uses is not introduced until 3.10. I can't get Python 3.9 to build the v1.6.4 docs because of this. My new recommendation is to either bump the minimum python version required or fix this by changing the semantics.

My current error is:
```bash
File "...\solar-data-tools\solardatatools\dataio.py", line 271, in <module>
    sensor: int | list[int] | None = None,
TypeError: unsupported operand type(s) for |: 'type' and 'types.GenericAlias'

mingw32-make: *** [Makefile:20: html] Error 2
```

Which would be fixed by:
```
from typing import Union, List

sensor: Union[int, List[int], None] = None
```
