### 2025-03-24 #210 @cdeline
Hi, I think I've completed my review.  Everything checks out for the most part - I opened a couple of issues that should be pretty simple to address, and there can never be too many pytests.  I'm happy to wait for incorporation of my comments, or not - they're pretty trivial.  Overall the software looks good and was pretty easy to navigate - admittedly I have loads of experience with time-series PV data, but I expect that someone else without as much experience would be able to get started with this package and could successfully follow the examples and get their own analysis up and running.
