### 2025-03-25 c6dd1c9
algorithm rewrite
