### 2025-03-25 596e3d2
variable renames and expose options to capacity change class method
