### 2025-03-25 d146d4a
enforce that tuned weight must be larger than the minimum holdout weight
