### 2025-03-26 010fb3c
update capacity clustering plot
