### 2025-03-26 974f0d6
adding constraint on large trends
