### 2025-03-26 68c2f5a
new weight optimization logic
