+++ solardatatools/algorithms/capacity_change.py
@@ -32,38 +32,57 @@ class CapacityChange:
         self.weight_vals = None
         self.holdout_error = None
         self.test_error = None
+        self.trend_costs = None
 
     def run(
         self,
         data,
+        metric=None,
         weight=None,
         filter=None,
         quantile=1.00,
         solver=None,
     ):
-        metric = np.nanquantile(data, q=quantile, axis=0)
-        metric_temp = np.ones_like(metric) * np.nan
-        # metric /= np.max(metric)
-        metric[~np.isnan(metric_temp)] = np.log(metric_temp[~np.isnan(metric_temp)])
+        if metric is None:
+            metric = np.nanquantile(data, q=quantile, axis=0)
+            metric_temp = np.ones_like(metric) * np.nan
+            # metric /= np.max(metric)
+            metric[~np.isnan(metric_temp)] = np.log(metric_temp[~np.isnan(metric_temp)])
         if filter is None:
             filter = ~np.isnan(metric)
         else:
             filter = np.logical_and(filter, ~np.isnan(metric))
         if weight is None:
-            weights = np.logspace(-1, 3, 13)
-            test_r, train_r, best_ix = self.optimize_weight(
-                metric, filter, weights, solver=solver
+            # initial check if changes are present
+            s1, s2, s3 = self.solve_sd(
+                metric=metric, filter=filter, weight=50, solver=solver
             )
-            tuned_weight = weights[best_ix]
+            rounded_s1 = custom_round(s1 - s1[0]) + s1[0]
+            set_labels = list(set(rounded_s1))
+            if len(set_labels) == 1:
+                tuned_weight = 50
+                best_ix = None
+                test_r = None
+                train_r = None
+                weights = None
+                trend_costs = None
+            else:
+                # changes are present, optimize the weight with holdout
+                weights = np.logspace(1, 3, 9)
+                test_r, train_r, best_ix, trend_costs = self.optimize_weight(
+                    metric, filter, weights, solver=solver
+                )
+                tuned_weight = weights[best_ix]
+                s1, s2, s3 = self.solve_sd(metric, filter, tuned_weight, solver=solver)
         else:
             tuned_weight = weight
             best_ix = None
             test_r = None
             train_r = None
             weights = None
-            changes_detected = None
-        # change detector, seasonal term, linear term
-        s1, s2, s3 = self.solve_sd(metric, filter, tuned_weight, solver=solver)
+            trend_costs = None
+            s1, s2, s3 = self.solve_sd(metric, filter, weight, solver=solver)
+
         # Get capacity assignments (label each day)
         # rounding buckets aligned to first value of component, bin widths of 0.05
         rounded_s1 = custom_round(s1 - s1[0]) + s1[0]
@@ -80,6 +99,7 @@ class CapacityChange:
         self.weight_vals = weights
         self.holdout_error = test_r
         self.train_error = train_r
+        self.trend_costs = trend_costs
 
     def solve_sd(
         self, metric, filter, weight, w3=1, w4=1e1, solver=None, verbose=False
@@ -106,6 +126,7 @@ class CapacityChange:
         # initialize results objects
         train_r = np.zeros_like(weights)
         test_r = np.zeros_like(weights)
+        trend_costs = np.zeros_like(weights)
         # iterate over possible values of weight parameter
         for i, v in enumerate(weights):
             s1, s2, s3 = self.solve_sd(
@@ -115,50 +136,14 @@ class CapacityChange:
             # collect results
             train_r[i] = np.average(np.abs((y - s1 - s2 - s3)[train]))
             test_r[i] = np.average(np.abs((y - s1 - s2 - s3)[test]))
-        # Precheck: Determine if changes are present. We observe the two conditions indicate a change:
[…]
