### 2025-03-28 b30440e
redesigned weight selection procedure and plotting function
