+++ solardatatools/algorithms/capacity_change.py
@@ -18,6 +18,7 @@ import numpy as np
 from solardatatools.signal_decompositions import l1_l1d1_l2d2p365
 from solardatatools.utilities import segment_diffs, make_pooled_dsig
 from sklearn.model_selection import train_test_split
+import matplotlib.pyplot as plt
 
 
 class CapacityChange:
@@ -32,7 +33,8 @@ class CapacityChange:
         self.weight_vals = None
         self.holdout_error = None
         self.test_error = None
-        self.trend_costs = None
+        self.jmp_counts = None
+        self.min_jumps = None
 
     def run(
         self,
@@ -46,41 +48,52 @@ class CapacityChange:
         if metric is None:
             metric = np.nanquantile(data, q=quantile, axis=0)
             metric_temp = np.ones_like(metric) * np.nan
-            # metric /= np.max(metric)
-            metric[~np.isnan(metric_temp)] = np.log(metric_temp[~np.isnan(metric_temp)])
+            slct = np.logical_and(~np.isnan(metric), metric > 0)
+            metric_temp[slct] = np.log(metric[slct])
+            metric = metric_temp
         if filter is None:
             filter = ~np.isnan(metric)
         else:
             filter = np.logical_and(filter, ~np.isnan(metric))
         if weight is None:
             # initial check if changes are present
-            s1, s2, s3 = self.solve_sd(
-                metric=metric, filter=filter, weight=50, solver=solver
+            # initial_weight = 50
+            # initial_weight = .1
+            # s1, s2, s3 = self.solve_sd(
+            #     metric=metric, filter=filter, weight=initial_weight, solver=solver
+            # )
+            # rounded_s1 = custom_round(s1 - s1[0]) + s1[0]
+            # set_labels = list(set(rounded_s1))
+            # if len(set_labels) == 1:
+            #     tuned_weight = initial_weight
+            #     best_ix = None
+            #     test_r = None
+            #     train_r = None
+            #     weights = None
+            #     jmp_counts = None
+            # else:
+            # changes are present, optimize the weight with holdout
+            # weights = np.logspace(1, 3, 9)
+            weights = np.logspace(-0.5, 2.5, 13)
+            test_r, train_r, best_ix, jmp_counts, min_jumps = self.optimize_weight(
+                metric, filter, weights, solver=solver
             )
-            rounded_s1 = custom_round(s1 - s1[0]) + s1[0]
-            set_labels = list(set(rounded_s1))
-            if len(set_labels) == 1:
-                tuned_weight = 50
-                best_ix = None
-                test_r = None
-                train_r = None
-                weights = None
-                trend_costs = None
-            else:
-                # changes are present, optimize the weight with holdout
-                weights = np.logspace(1, 3, 9)
-                test_r, train_r, best_ix, trend_costs = self.optimize_weight(
-                    metric, filter, weights, solver=solver
-                )
-                tuned_weight = weights[best_ix]
-                s1, s2, s3 = self.solve_sd(metric, filter, tuned_weight, solver=solver)
+            tuned_weight = weights[best_ix]
+            s1, s2, s3 = self.solve_sd(metric, filter, tuned_weight, solver=solver)
+            num_jumps = np.sum(~np.isclose(np.diff(s1), 0))
+            num_years = np.max([1, len(metric) / 365])
+            # if num_jumps > num_years * 4:
+            #     tuned_weight = weights[-1]
+            #     best_ix = len(weights) - 1
+            #     s1, s2, s3 = self.solve_sd(metric, filter, tuned_weight, solver=solver)
         else:
             tuned_weight = weight
             best_ix = None
[…]
