### 2025-03-28 ecf153c
Refactoring to match other signal decomposition functions
