+++ solardatatools/signal_decompositions.py
@@ -28,20 +28,15 @@ SDT algorithms. The defined signal decompositions are:
    - constrained to have first val at 0 and last val at 1
 
 """
-import sys
-import numpy as np
-import cvxpy as cvx
-from spcqe import make_basis_matrix, make_regularization_matrix
 
 from solardatatools._osd_signal_decompositions import (
     _osd_l2_l1d1_l2d2p365,
-    _osd_l1_l1d1_l2d2p365,
     _osd_tl1_l2d2p365,
     _osd_l2_l1d2_constrained,
 )
 from solardatatools._cvx_signal_decompositions import (
     _cvx_l2_l1d1_l2d2p365,
-    _cvx_l1_l1d1_l2d2p365,
+    _cvx_l1_pwc_smoothper_trend,
     _cvx_tl1_l2d2p365,
     _cvx_l2_l1d2_constrained,
 )
@@ -79,8 +74,7 @@ def l2_l1d1_l2d2p365(
         seasonal signal
     :param yearly_periodic: Adds periodicity constraint to signal decomposition
     :param return_all: Returns all components and the objective value. Used for tests.
-    :param solver: Solver to use for the decomposition. QSS and OSQP are supported with
-        OSD. MOSEK will trigger CVXPY use.
+    :param solver: Solver to use for the decomposition. Supported solvers are CLARABEL and MOSEK.
     :param sum_card: Boolean for using the nonconvex formulation using the cardinality penalty,
         Supported only using OSD with the QSS solver.
     :param transition_locs: List of indices where transitions are located. Only used in CVXPY problems.
@@ -148,7 +142,7 @@ def tl1_l2d2p365(
     solardatatools/sunrise_sunset.py
 
 
-    This is a convex problem and the default solver across SDT is OSQP.
+    This is a convex problem and the default solver across SDT is CLARABEL.
 
     :param signal: A 1d numpy array (must support boolean indexing) containing
         the signal of interest
@@ -162,8 +156,7 @@ def tl1_l2d2p365(
         seasonal signal
     :param yearly_periodic: Adds periodicity constraint to signal decomposition
     :param return_all: Returns all components and the objective value. Used for tests.
-    :param solver: Solver to use for the decomposition. QSS and OSQP are supported with
-        OSD. MOSEK will trigger CVXPY use.
+    :param solver: Solver to use for the decomposition. Supported solvers are CLARABEL and MOSEK.
     :param verbose: Sets verbosity
     :return: A tuple with three 1d numpy arrays containing the three signal component estimates
     """
@@ -192,7 +185,7 @@ def tl1_l2d2p365(
     return res
 
 
-def l1_l1d1_l2d2p365(
+def l1_pwc_smoothper_trend(
     signal,
     use_ixs=None,
     w2=2e1,
@@ -216,69 +209,21 @@ def l1_l1d1_l2d2p365(
     :param w2: Weight on the piecewise constant component
     :param w3: Weight on the smooth, periodic component
     :param w4: Weight on the slope of the trend term (discourages large trends)
-    :param solver: Solver to use for the decomposition. Standard cvxpy solvers are supported
+    :param solver: Solver to use for the decomposition. Supported solvers are CLARABEL and MOSEK.
     :param verbose: Sets verbosity
     :return: A tuple with three 1d numpy arrays containing the three non-noise signal component estimates
     """
-    if solver is not None:
-        if solver.upper() not in ["MOSEK", "CLARABEL", "OSQP"]:
-            solver = "CLARABEL"
[…]
