+++ solardatatools/_cvx_signal_decompositions.py
@@ -23,20 +23,22 @@ This module contains standardized signal decomposition models for use in the SDT
    - Constrained to have first value at 0 and last value at 1
 
 """
+
 import sys
 import numpy as np
 
 import cvxpy as cvx
+from spcqe import make_basis_matrix, make_regularization_matrix
 
 
 def _cvx_l2_l1d1_l2d2p365(
     signal,
     use_ixs=None,
-    w0=10, # "hard-coded"
-    w1=50, # optimized
+    w0=10,  # "hard-coded"
+    w1=50,  # optimized
     w2=1e5,
     yearly_periodic=False,
-    return_all=False, 
+    return_all=False,
     solver="MOSEK",
     transition_locs=None,
     verbose=False,
@@ -98,7 +100,7 @@ def _cvx_l2_l1d1_l2d2p365(
         s_seas = cvx.Variable(len(signal))
         s_error = cvx.Variable(len(signal))
 
-        if transition_locs is None: # TODO: this should be two separate sd problems
+        if transition_locs is None:  # TODO: this should be two separate sd problems
             objective = cvx.Minimize(
                 w0 * cvx.sum_squares(s_error)
                 + w1 * cvx.norm1(cvx.multiply(tv_weights, cvx.diff(s_hat, k=1)))
@@ -106,12 +108,12 @@ def _cvx_l2_l1d1_l2d2p365(
             )
         else:
             objective = cvx.Minimize(
-                w0 * cvx.norm(s_error)
-               + w2 * cvx.sum_squares(cvx.diff(s_seas, k=2))
+                w0 * cvx.norm(s_error) + w2 * cvx.sum_squares(cvx.diff(s_seas, k=2))
             )
         # Consistency constraints
         constraints = [
-            signal[index_set] == s_hat[index_set] + s_seas[index_set] + s_error[index_set],
+            signal[index_set]
+            == s_hat[index_set] + s_seas[index_set] + s_error[index_set],
             cvx.sum(s_seas[:365]) == 0,
         ]
         if len(signal) > 365:
@@ -137,13 +139,13 @@ def _cvx_l2_l1d1_l2d2p365(
 def _cvx_tl1_l2d2p365(
     signal,
     use_ixs=None,
-    tau=0.75, # passed as 0.05/0.1 (sunrise), 0.95/0.9 (sunset)
+    tau=0.75,  # passed as 0.05/0.1 (sunrise), 0.95/0.9 (sunset)
     w0=1,
     w1=500,
     yearly_periodic=True,
     return_all=False,
     solver="MOSEK",
-    verbose=False
+    verbose=False,
 ):
     """
     Used in:
@@ -173,8 +175,8 @@ def _cvx_tl1_l2d2p365(
     x = cvx.Variable(len(signal))
     r = signal[use_ixs] - x[use_ixs]
     objective = cvx.Minimize(
-        w0 * cvx.sum(0.5 * cvx.abs(r) + (tau - 0.5) * r) +
-        w1 * cvx.sum_squares(cvx.diff(x, k=2))
+        w0 * cvx.sum(0.5 * cvx.abs(r) + (tau - 0.5) * r)
+        + w1 * cvx.sum_squares(cvx.diff(x, k=2))
     )
     if len(signal) > 365 and yearly_periodic:
         constraints = [x[365:] == x[:-365]]
@@ -189,88 +191,106 @@ def _cvx_tl1_l2d2p365(
     return x.value
 
 
-def _cvx_l1_l1d1_l2d2p365(
+def _cvx_l1_pwc_smoothper_trend(
     signal,
     use_ixs=None,
-    w0=3,  # l1 term
-    w1=18, # l1d1 term
[…]
