### 2025-04-06 slacgismo/solar-data-tools#194 @echedey-ls
[pyOpenSci reviewer] ReadTheDocs version picker is broken
**Issue**
1. Go to https://solar-data-tools.readthedocs.io/en/main/
2. Click on _Choose version_ (upper right corner)
3. See no options are shown

![Image](https://github.com/user-attachments/assets/560c6df8-2109-447a-9fed-2906d98f1f64)

**Suggestion**
Fix the version picker.

I can't tell how to do so cause I've never used it.

Obviously, these is yet another minor issue that won't get in the way of a favourable endorsement, but would be great to fix.
