### 2025-04-06 slacgismo/solar-data-tools#195 @echedey-ls
[pyOpenSci reviewer] Do not raise bare Exceptions
**Issue**
Currently, codebase makes use of `raise Exception(...)` in quite some places.
It's advised not to raise this general exception and narrow the type that is raised, so it's easier to filter in try except clauses and also gives a slightly more info to the user.

Instances of this statement are in:

```bash
\solar-data-tools> git grep -n "raise Exception("
sdt_dask/clients/local_client.py:74:            raise Exception(f"workers and threads exceed local resources, {self.cpu_count} cores present")
solardatatools/algorithms/loss_factor_analysis.py:590:        raise Exception("destination must be larger than origin in all dimensions")
solardatatools/data_handler.py:122:                    raise Exception(e)
solardatatools/dataio.py:372:            raise Exception(
solardatatools/dataio.py:408:            raise Exception(
solardatatools/dataio.py:422:                raise Exception("Empty dataframe returned from query")
solardatatools/dataio.py:450:        raise Exception("Failed to get query info")
solardatatools/dataio.py:504:        raise Exception("Empty dataframe returned from query")
solardatatools/polar_transform.py:42:                raise Exception(e)
```

**Suggestion**
I haven't checked in depth all the `git grep` results, but some of them do look like they can be replaced by `RuntimeException` (sdt_dask/clients/local_client.py:74), `ValueError` (solardatatools/data_handler.py:122 & solardatatools/polar_transform.py:42) or `IOError` (solardatatools/dataio.py).

You may as well define your custom Exceptions.
