### 2025-04-18 47bcb5b
Make the default behavior to load a standard CSV file and not apply make_timeseries subroutine.
