### 2025-04-18 b5aa44e
assume standard schema with timestamps in first column
