### 2025-04-18 1764b9a
drop support for python 3.9 and clean up pyproject.toml
