+++ pyproject.toml
@@ -14,7 +14,7 @@ maintainers = [
   { name = "Sara Miskovich", email = "smiskov@slac.stanford.edu" },
 ]
 keywords = ["solar", "pv", "photovoltaic"]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 
 dependencies = [
     "scipy",
@@ -41,10 +41,6 @@ classifiers = [
     "Development Status :: 4 - Beta",
     "License :: OSI Approved :: BSD License",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.6",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
