### 2025-04-18 slacgismo/solar-data-tools#186 @pluflou
Thank you @echedey-ls for this thorough submission! I agree that this is kind of messy. Dask gets weird when versions don't match exactly so we ended with this mix of requirements to get people set up locally, on the cloud and using Docker. I will work on getting this cleaned up.
