### 2025-04-18 21b147c
clarify timeshift analysis printed error
