+++ solardatatools/data_handler.py
@@ -2322,7 +2322,10 @@ time zone errors     {report['time zone correction'] != 0}
             fig = plt.gcf()
             return fig
         else:
-            print("Please run pipeline first.")
+            print(
+                "Please run the time shift analysis first by setting 'fix_shifts=True' when "
+                "running the pipeline, e.g. 'DataHandler.run_pipeline(fix_shifts=True)'."
+            )
 
     def plot_circ_dist(self, flag="good", num_bins=12 * 4, figsize=(8, 8)):
         """
