### 2025-04-18 3bf160c
clarify timeshift analysis printed error
