### 2025-04-30 #210 @echedey-ls
Review done! I know I've left a good amount of comments, but that does not mean they all are blockers. Except for:
1. https://github.com/slacgismo/solar-data-tools/issues/185 and
2. https://github.com/slacgismo/solar-data-tools/issues/184,
that I find important for end users.

I haven't had much trouble understanding the repo, thou it's kind of big at first with archived notebooks, then notebooks on the web, etc. I, unlike @cdeline , don't have much experience with time-series of data analysis, but found the concepts and use of code easy enough to understand with my novel background in PV. Well done.

@cmarmo feel free to guide me through next steps if something feels off or further review of specific sections are deemed necessary. @pluflou , if you need any clarification or helping hand with something, I may be able to assist.

P.D.: I'll keep an eye on the PRs for some time and leave comments where I find important.

(EDIT: I don't ask for the enforcement of the [two-week feedback rule](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#review-follow-up), mostly cause I'm the one unable to follow it 😬 )
