### 2025-04-30 slacgismo/solar-data-tools#186 @echedey-ls
Feel free to ask for help or review. For clarification, I think the manageable decision would be to either keep `docker/requirements.txt` or `pyproject.toml` optional dependency group `[dask]`. In any case, I find potentially useful and trouble-free to create the image based on the current repo contents and not the latest published SDT version.
