### 2025-05-01 #210 @jinningwang
Thanks for the efforts in this work! My review report is here.

Readme file**  **requirements

**    **- [x] Continuous integration and test coverage,

**    **- [x] Docs building (if you have a documentation website),

**    **- [x] A [repostatus.org](https://www.repostatus.org/) badge,

**    **- [x] Python versions supported,

**    **- [x] Current package version (on PyPI / Conda).

**    **- [x] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)

**  **- [x] All tests pass on the reviewer's local machine for the package version submitted by the author. Ideally this should be a tagged version making it easy for reviewers to install.

**  **- [x] Tests cover essential functions of the package and a reasonable range of inputs and conditions.

**    **A few notable highlights to look at:

**    **- [x] Package supports modern versions of Python and not [End of life versions](https://endoflife.date/python).

**    **- [ ] Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)

The package contains a `paper.md` matching [JOSS&#39;s requirements](http://joss.theoj.org/about#paper_structure) with:

2.5
---

The package is well structured and easy to use.
The documentation is clear and provides good examples of how to use the package.
The installation process is straightforward and works as documented.
Some concerns are listed as below:
1. A list of supported Python versions would be helpful in the README file.
1. Formatting issues are present in the codebase. My local test is excerpted at the end of this review for your reference.
1. A `repostatus.org` badge in the README file might be useful to indicate the status of the package.

Additional info is provided below:
My local test environment is created via conda with following packages:
- Python 3.12.10
- numpy 2.2.5
- pandas    2.2.3
- pytest    8.3.5
- flake8    7.2.0
- solar-data-tools  1.7.1

Testing lint using flake8 on commit ``18e2939bcf673eb2c234353e142a9152b7db4c66``, shows some formatting issues.
Some of them are:
./solardatatools/clear_day_detection.py:60:80: E501 line too long (105 > 79 characters)
./solardatatools/clear_day_detection.py:61:80: E501 line too long (99 > 79 characters)
./solardatatools/clear_day_detection.py:63:80: E501 line too long (118 > 79 characters)
./solardatatools/clear_day_detection.py:65:80: E501 line too long (81 > 79 characters)
./solardatatools/clear_day_detection.py:71:80: E501 line too long (83 > 79 characters)
./solardatatools/clear_day_detection.py:72:80: E501 line too long (81 > 79 characters)
./solardatatools/clear_day_detection.py:78:80: E501 line too long (80 > 79 characters)
./solardatatools/clear_day_detection.py:88:80: E501 line too long (111 > 79 characters)
./solardatatools/clear_day_detection.py:96:80: E501 line too long (84 > 79 characters)
./solardatatools/clear_day_detection.py:98:80: E501 line too long (91 > 79 characters)
./solardatatools/clear_time_labeling.py:30:80: E501 line too long (83 > 79 characters)
./solardatatools/clear_time_labeling.py:32:80: E501 line too long (82 > 79 characters)
./solardatatools/clear_time_labeling.py:40:80: E501 line too long (83 > 79 characters)
./solardatatools/clear_time_labeling.py:42:80: E501 line too long (85 > 79 characters)
./solardatatools/clear_time_labeling.py:43:80: E501 line too long (81 > 79 characters)
./solardatatools/data_filling.py:10:1: F401 'matplotlib.pyplot as plt' imported but unused
./solardatatools/data_filling.py:11:1: F401 'solardatatools.daytime.find_daytime' imported but unused
./solardatatools/data_handler.py:17:17: E401 multiple imports on one line
./solardatatools/data_handler.py:46:1: E402 module level import not at top of file
```
