### 2025-05-05 e7a6d24
attempt to fix version switcher in RTD
