### 2025-05-05 37bf34f
fix install in test pipeline
