### 2025-05-05 slacgismo/solar-data-tools#194 @pluflou
Since Read the Docs has their own switcher on the bottom right, I am doing to disable this button in the Sphinx theme.
