### 2025-05-05 ac06374
remove sphinx version switcher in favor of RTD one
