### 2025-05-23 ed81655
add python versions to readme badges
