+++ README.md
@@ -17,6 +17,13 @@
 
 <table  align="center" >
 <tr>
+  <td>Repo Status</td>
+    <td>
+    <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" 
+      alt="Project Status: Active – The project has reached a stable, usable state and is being actively developed." />
+    </a>
+  </tr>
+  <tr>
   <td>Python Versions</td>
   <td>
     <a href="https://www.python.org/downloads/release/python-3100/">
