+++ pvsystemprofiler/scripts/parameter_estimation_script.py
@@ -1,9 +1,10 @@
-""" Longitude run script
+"""Longitude run script
 This run script allows to run the longitude_study for multiple sites. The site ids to be evaluated can be provided in
  a csv file. Alternatively, the path to a folder containing the input signals of the sites in separate csv files can be
  provided.  The script provides the option to provided the full path to csv file containing latitude and gmt offset for
  each system for comparison.
 """
+
 import sys
 from pathlib import Path
 import pandas as pd
@@ -23,108 +24,165 @@ from pvsystemprofiler.scripts.modules.script_functions import generate_list
 from solardatatools.dataio import load_cassandra_data
 from pvsystemprofiler.scripts.modules.script_functions import extract_sys_parameters
 from pvsystemprofiler.scripts.modules.script_functions import get_commandline_inputs
-from pvsystemprofiler.scripts.modules.script_functions import run_failsafe_lon_estimation
-from pvsystemprofiler.scripts.modules.script_functions import run_failsafe_lat_estimation
+from pvsystemprofiler.scripts.modules.script_functions import (
+    run_failsafe_lon_estimation,
+)
+from pvsystemprofiler.scripts.modules.script_functions import (
+    run_failsafe_lat_estimation,
+)
 from pvsystemprofiler.scripts.modules.script_functions import run_failsafe_ta_estimation
 from solardatatools import DataHandler
 
 
 def evaluate_systems(site_id, inputs_dict, df, site_metadata, json_file_dict=None):
-    partial_df_cols = ['site', 'system', 'passes pipeline', 'length', 'capacity_estimate', 'data_sampling',
-                       'data quality_score', 'data clearness_score', 'inverter_clipping', 'time_shifts_corrected',
-                       'time_zone_correction', 'capacity_changes', 'normal_quality_scores']
+    partial_df_cols = [
+        "site",
+        "system",
+        "passes pipeline",
+        "length",
+        "capacity_estimate",
+        "data_sampling",
+        "data quality_score",
+        "data clearness_score",
+        "inverter_clipping",
+        "time_shifts_corrected",
+        "time_zone_correction",
+        "capacity_changes",
+        "normal_quality_scores",
+    ]
 
     if json_file_dict is not None:
-        partial_df_cols.extend(['zip_code', 'real longitude', 'real latitude', 'real tilt', 'real azimuth'])
-    if inputs_dict['time_shift_manual']:
-        partial_df_cols.append('time_shift_manual')
+        partial_df_cols.extend(
+            ["zip_code", "real longitude", "real latitude", "real tilt", "real azimuth"]
+        )
+    if inputs_dict["time_shift_manual"]:
+        partial_df_cols.append("time_shift_manual")
 
     partial_df = pd.DataFrame(columns=partial_df_cols)
 
-    ll = len(inputs_dict['power_column_label'])
+    ll = len(inputs_dict["power_column_label"])
 
-    if inputs_dict['convert_to_ts']:
-        dh = DataHandler(df, convert_to_ts=inputs_dict['convert_to_ts'])
+    if inputs_dict["convert_to_ts"]:
+        dh = DataHandler(df, convert_to_ts=inputs_dict["convert_to_ts"])
         cols = [el[-1] for el in dh.keys]
     else:
         cols = df.columns
 
     i = 0
     for col_label in cols:
-        if col_label.find(inputs_dict['power_column_label']) != -1:
+        if col_label.find(inputs_dict["power_column_label"]) != -1:
             system_id = col_label[ll:]
-            if system_id in site_metadata['system'].tolist():
+            if system_id in site_metadata["system"].tolist():
                 i += 1
-                dh = DataHandler(df, convert_to_ts=inputs_dict['convert_to_ts'])
[…]
