+++ pvsystemprofiler/latitude_study.py
@@ -1,4 +1,4 @@
-""" Latitude Study Module
+"""Latitude Study Module
 This module contains a class for conducting a study to estimate latitude from solar data. This code accepts solar data
 in the form of a `solar-data-tools` `DataHandler` object, which is used to standardize and pre-process the data. The
 provided class will then estimate the latitude of the site that produced the data, using the `run` method.
@@ -11,17 +11,22 @@ The following configurations can be run:
  - Day selection method: 'all', 'clear', 'cloudy'
 
 """
+
 import numpy as np
 import pandas as pd
 from pvsystemprofiler.utilities.declination_equation import delta_spencer
 from pvsystemprofiler.utilities.declination_equation import delta_cooper
 from pvsystemprofiler.algorithms.latitude.hours_daylight import calculate_hours_daylight
-from pvsystemprofiler.algorithms.latitude.hours_daylight import calculate_hours_daylight_raw
-from pvsystemprofiler.algorithms.optimized_sunrise_sunset import get_optimized_sunrise_sunset
+from pvsystemprofiler.algorithms.latitude.hours_daylight import (
+    calculate_hours_daylight_raw,
+)
+from pvsystemprofiler.algorithms.optimized_sunrise_sunset import (
+    get_optimized_sunrise_sunset,
+)
 from pvsystemprofiler.algorithms.latitude.estimation import estimate_latitude
 
 
-class LatitudeStudy():
+class LatitudeStudy:
     def __init__(self, data_handler, lat_true_value=None):
         """
         :param data_handler: `DataHandler` class instance loaded with a solar power data set.
@@ -31,7 +36,7 @@ class LatitudeStudy():
         self.data_handler = data_handler
         self.latitude_true_value = lat_true_value
         if not data_handler._ran_pipeline:
-            print('Running DataHandler preprocessing pipeline with defaults')
+            print("Running DataHandler preprocessing pipeline with defaults")
             self.data_handler.run_pipeline()
         self.data_matrix = self.data_handler.filled_data_matrix
         self.raw_data_matrix = self.data_handler.raw_data_matrix
@@ -61,10 +66,19 @@ class LatitudeStudy():
         # Results
         self.results = None
 
-    def run(self, data_matrix=('raw', 'filled'),
-            daylight_method=('raw daylight', 'sunrise-sunset', 'optimized_estimates', 'optimized_measurements'),
-            delta_method=('cooper', 'spencer'), day_selection_method=('all', 'clear', 'cloudy'),
-            threshold=None):
+    def run(
+        self,
+        data_matrix=("raw", "filled"),
+        daylight_method=(
+            "raw daylight",
+            "sunrise-sunset",
+            "optimized_estimates",
+            "optimized_measurements",
+        ),
+        delta_method=("cooper", "spencer"),
+        day_selection_method=("all", "clear", "cloudy"),
+        threshold=None,
+    ):
         """
         Run a study with the given configuration of options. Defaults to
         running all available options. Any kwarg can be constrained by
@@ -95,100 +109,140 @@ class LatitudeStudy():
         day_selection_method = np.atleast_1d(day_selection_method)
 
         if threshold is None:
-            self.daytime_threshold = 0.001 * np.ones(len(data_matrix) * len(daylight_method) * len(delta_method) *
-                                                     len(day_selection_method))
+            self.daytime_threshold = 0.001 * np.ones(
+                len(data_matrix)
+                * len(daylight_method)
+                * len(delta_method)
+                * len(day_selection_method)
+            )
         else:
             self.daytime_threshold = threshold
 
         self.delta_cooper = delta_cooper(self.day_of_year, self.daily_meas)
[…]
