### 2025-05-23 6e023df
adjust all bare exceptions
