+++ .github/workflows/lint.yml
@@ -1,0 +1,41 @@
+name: Lint
+on:
+  pull_request:
+  workflow_dispatch:
+  push:
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 1
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+
+      - name: Install SDT
+        run: |
+          pip install -e ".[dev]"
+
+      - name: Install pre-commit
+        run: pip install pre-commit
+
+      - name: List Python package versions
+        run: pip list
+
+      - name: Switch to a temporary branch
+        run: |
+          git checkout -b _ci_branch_
+
+      - name: Run pre-commit on all files
+        env:
+          RUFF_OUTPUT_FORMAT: github
+        run: |
+          pre-commit run \
+            --show-diff-on-failure \
+            --color=always \
+            --all-files
