+++ statistical_clear_sky/algorithm/plot/plot_mixin.py
@@ -1,207 +1,0 @@
-"""
-This module defines Mixin for plot for IterativeClearSky.
-"""
-
-import numpy as np
-import seaborn as sns
-import matplotlib.pyplot as plt
-from solardatatools import plot_2d
-
-
-class PlotMixin(object):
-    def plot_lr(self, figsize=(14, 10), show_days=False):
-        fig, ax = plt.subplots(nrows=2, ncols=2, figsize=figsize)
-        ax[0, 1].plot(self._r_cs_value[0])
-        ax[1, 1].plot(self._r_cs_value[1:].T)
-        ax[0, 0].plot(self._l_cs_value[:, 0])
-        ax[1, 0].plot(self._l_cs_value[:, 1:])
-        ax[0, 0].legend(["$\\ell_1$"])
-        ax[1, 0].legend(
-            ["$\\ell_{}$".format(ix) for ix in range(2, self._r_cs_value.shape[0] + 1)]
-        )
-        ax[0, 1].legend(["$r_{1}$"])
-        ax[1, 1].legend(
-            ["$r_{}$".format(ix) for ix in range(2, self._r_cs_value.shape[0] + 1)]
-        )
-        if show_days:
-            use_day = self._obtain_weights_for_plotting() > 1e-1
-            days = np.arange(self._power_signals_d.shape[1])
-            ax[0, 1].scatter(
-                days[use_day], self._r_cs_value[0][use_day], color="orange", alpha=0.7
-            )
-        plt.tight_layout()
-        return fig
-
-    def plot_energy(
-        self, figsize=(12, 6), show_days=True, show_clear=True, scale_power=False
-    ):
-        if scale_power:
-            c = 1.0 / 1000
-        else:
-            c = 1.0
-        plt.figure(figsize=figsize)
-        plt.plot(
-            np.sum(self._power_signals_d, axis=0)
-            * 24
-            * c
-            / self._power_signals_d.shape[0],
-            linewidth=1,
-            alpha=0.7,
-            label="measured daily energy",
-        )
-        if show_clear:
-            plt.plot(
-                (self._r_cs_value[0] * np.sum(self._l_cs_value[:, 0]))
-                * 24
-                * c
-                / self._power_signals_d.shape[0],
-                linewidth=1,
-                label="estimated clear sky model",
-            )
-        if show_days:
-            use_day = self._obtain_weights_for_plotting() > 1e-1
-            days = np.arange(self._power_signals_d.shape[1])
-            plt.scatter(
-                days[use_day],
-                np.sum(self._power_signals_d, axis=0)[use_day]
-                * 24
-                * c
-                / self._power_signals_d.shape[0],
-                color="orange",
-                alpha=0.7,
-                label="detected clear days",
-            )
-        plt.legend()
-        plt.title("Daily Energy Signal")
-        fig = plt.gcf()
-        return fig
-
-    def plot_singular_vectors(self, k=4, figsize=(10, 4), show_days=False):
-        fig, ax = plt.subplots(nrows=k, ncols=2, figsize=(figsize[0], 2 * figsize[1]))
-        for i in range(k):
-            ax[i][0].plot(self._matrix_l0.T[i], linewidth=1)
-            ax[i][0].set_xlim(0, self._power_signals_d.shape[0])
-            ax[i][0].set_ylabel("$\\ell_{}$".format(i + 1))
-            ax[i][1].plot(self._matrix_r0[i], linewidth=1)
-            ax[i][1].set_xlim(0, self._power_signals_d.shape[1])
-            ax[i][1].set_ylabel("$r_{}$".format(i + 1))
-        ax[-1][0].set_xlabel("$i \\in 1, \\ldots, m$ (5-minute periods in one day)")
[…]
