### 2025-05-23 b29173d
skip checking notebooks, more linting
