+++ .pre-commit-config.yaml
@@ -19,4 +19,6 @@ repos:
     hooks:
       - id: ruff
         args: [--fix]
+        exclude: ".*\\.ipynb$"
       - id: ruff-format
+        exclude: ".*\\.ipynb$"
