### 2025-05-23 97c765f
add docs about pre-commit/linting requirements
