### 2025-05-23 c24165e
remove scs and associated things
