+++ statistical_clear_sky/algorithm/iterative_fitting.py
@@ -1,1010 +1,0 @@
-"""
-This module defines "Statistical Clear Sky Fitting" algorithm.
-"""
-
-from time import time
-import numpy as np
-from numpy.linalg import norm
-import cvxpy as cvx
-from collections import defaultdict
-from statistical_clear_sky.algorithm.initialization.singular_value_decomposition import (
-    SingularValueDecomposition,
-)
-from statistical_clear_sky.algorithm.initialization.linearization_helper import (
-    LinearizationHelper,
-)
-from statistical_clear_sky.algorithm.initialization.weight_setting import WeightSetting
-from statistical_clear_sky.algorithm.exception import ProblemStatusError
-from statistical_clear_sky.algorithm.minimization.left_matrix import (
-    LeftMatrixMinimization,
-)
-from statistical_clear_sky.algorithm.minimization.right_matrix import (
-    RightMatrixMinimization,
-)
-from statistical_clear_sky.algorithm.serialization.state_data import StateData
-from statistical_clear_sky.algorithm.serialization.serialization_mixin import (
-    SerializationMixin,
-)
-from statistical_clear_sky.algorithm.plot.plot_mixin import PlotMixin
-from statistical_clear_sky.utilities.data_loading import resample_index
-from statistical_clear_sky.utilities.progress import progress
-
-
-class IterativeFitting(SerializationMixin, PlotMixin):
-    """
-    Implementation of "Statistical Clear Sky Fitting" algorithm.
-    """
-
-    def __init__(
-        self,
-        data_matrix=None,
-        data_handler_obj=None,
-        rank_k=6,
-        solver_type="MOSEK",
-        reserve_test_data=False,
-    ):
-        """
-
-        :param data_matrix:
-        :param data_handler_obj:
-        :param rank_k:
-        :param solver_type:
-        :param reserve_test_data:
-        """
-        self._solver_type = solver_type
-        self._rank_k = rank_k
-        if data_handler_obj is None and data_matrix is None:
-            print("Please initialize class with a data set")
-        elif data_handler_obj is not None:
-            data_matrix = data_handler_obj.filled_data_matrix
-            self._power_signals_d = data_matrix
-            self._capacity = data_handler_obj.capacity_estimate
-            # Set the weighting now, to use the error flagging feature
-            weights = self._get_weight_setting().obtain_weights(data_matrix)
-            weights *= data_handler_obj.daily_flags.no_errors
-        else:
-            self._power_signals_d = data_matrix
-
-        self._decomposition = SingularValueDecomposition()
-        self._decomposition.decompose(data_matrix, rank_k=rank_k)
-
-        self._matrix_l0 = self._decomposition.matrix_l0
-        self._matrix_r0 = self._decomposition.matrix_r0
-        self._bootstrap_samples = None
-
-        self._set_testdays(data_matrix, reserve_test_data)
-        # Handle both DataHandler objects and reserving test data
-        if data_handler_obj is not None and self._test_days is not None:
-            weights[self._test_days] = 0
-
-        # Stores the current state of the object:
-        self._state_data = StateData()
-        self._store_initial_state_data()
-        if data_handler_obj is not None:
-            self._weights = weights
-            self._state_data.weights = weights
-
-        self._set_residuals()
-
-    def execute(
-        self,
-        mu_l=None,
-        mu_r=None,
-        tau=None,
-        exit_criterion_epsilon=1e-3,
-        max_iteration=10,
-        is_degradation_calculated=True,
-        max_degradation=None,
-        min_degradation=None,
-        non_neg_constraints=False,
-        verbose=True,
-        bootstraps=None,
-    ):
-
[…]
