### 2025-05-23 e2fb6c0
adjust typo in install for tests
