### 2025-05-23 7091e46
adjust docs regarding timezones/shifts
