### 2025-05-23 slacgismo/solar-data-tools#201 @pluflou
DOC: Remove duplicate notebooks
Adjust any duplicated notebooks between `notebooks/examples` and `docs/source/getting_started/notebooks` to reference the notebooks at `/notebooks` from docs with a relative path.

For example: `notebooks\examples\DataHandler_DemoDefault.ipynb` vs `docs\source\getting_started\notebooks\demo_default.ipynb` (the latter has at least one plot at the end the former does not)
