### 2025-05-23 slacgismo/solar-data-tools#202 @pluflou
ENH: put packages under once namespace and move to src/ repo layout
For ease of use, consider putting all packages distributed under one same namespace (so they would instead become modules of a common package), for convention with most Python packages.

Also consider moving to src/ layout, benefits detailed [here](https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/), e.g.,
* The src layout requires installation of the project to be able to run its code, and the flat layout does not.
* The src layout helps prevent accidental usage of the in-development copy of the code.
