### 2025-05-23 slacgismo/solar-data-tools#203 @pluflou
ENH: <Please write a comprehensive title after the 'ENH: ' prefix>
Add an optional `ax` parameter to `plot_*` functions. It would ease the creation of a complex data analysis for many sites, for example, by easing the generation of multi-axes plots, or web-based services.

The main plotting functions to add this to would be:
- DataHandler.plot_heatmap
- DataHandler.plot_circ_dist
- DataHandler.plot_daily_energy
- DataHandler.plot_daily_signals
- DataHandler.plot_density_signal
- DataHandler.plot_data_quality_scatter
- DataHandler.plot_capacity_change_analysis
- DataHandler.plot_time_shift_analysis_results
- DataHandler.plot_clipping
- DataHandler.plot_cdf_analysis
- DataHandler.plot_daily_max_cdf_and_pdf
- DataHandler.plot_polar_transform
