### 2025-05-23 #210 @pluflou
Hi all,

Many thanks to @cdeline, @echedey-ls and @jinningwang for their reviews. I have created three PRs that address the reviewers' comments, as well as Issues that note things that were raised that weren't gating for this review for future work. Please feel free to review/comment as needed.

**Open PRs**
- slacgismo/solar-data-tools#197 contains most of the requested changes and closes all of the related open issues
- slacgismo/solar-data-tools#198 follows @jinningwang's suggestion to implement linting
- slacgismo/solar-data-tools#200 isn't directly related to the comments, but it removes the deprecated StatisticalClearSky package, and I plan to merge this one first and update PRs 197 and 198 accordingly

**Open Issues (for future work)**
- https://github.com/slacgismo/solar-data-tools/issues/201
- https://github.com/slacgismo/solar-data-tools/issues/202
- https://github.com/slacgismo/solar-data-tools/issues/203

If I missed anything, please let me know, otherwise, I'll wait for comments/approvals/confirmation before we merge the PRs. Thank you all!
