### 2025-06-01 #210 @pluflou
@cmarmo ok great, that sounds reasonable to me! I'll merge it tomorrow and ping everyone when I'm done. Thank you!
