### 2025-06-02 slacgismo/solar-data-tools#186 @echedey-ls
Looks good to me, thanks for the in-depth explanation!
