+++ solardatatools/data_handler.py
@@ -5,8 +5,6 @@ This module contains a class for managing a data processing pipeline
 
 """
 
-# ruff: noqa: E731, E722
-
 from time import time
 from datetime import timedelta
 from datetime import datetime
@@ -19,7 +17,6 @@ import matplotlib.cm as cm
 import traceback
 import sys
 from tqdm import tqdm
-from pandas.plotting import register_matplotlib_converters
 from solardatatools.time_axis_manipulation import (
     make_time_series,
     standardize_time_axis,
@@ -35,7 +32,9 @@ from solardatatools.data_quality import (
 from solardatatools.data_filling import zero_nighttime, interp_missing
 from solardatatools.clear_day_detection import ClearDayDetection
 from solardatatools.plotting import plot_2d
+from solardatatools.clear_time_labeling import find_clear_times
 from solardatatools.solar_noon import avg_sunrise_sunset
+from solardatatools.polar_transform import PolarTransform
 from solardatatools.algorithms import (
     CapacityChange,
     TimeShift,
@@ -43,9 +42,7 @@ from solardatatools.algorithms import (
     ClippingDetection,
     LossFactorAnalysis,
 )
-
-from solardatatools.polar_transform import PolarTransform
-
+from pandas.plotting import register_matplotlib_converters
 
 register_matplotlib_converters()
 
@@ -195,6 +192,7 @@ class DataHandler:
         # Useful daily signals defined by the data set
         self.daily_signals = DailySignals()
         # Algorithm objects
+        self.scsf = None
         self.capacity_analysis = None
         self.time_shift_analysis = None
         self.daytime_analysis = None
@@ -436,7 +434,7 @@ class DataHandler:
         try:
             ss.run_optimizer(self.raw_data_matrix, plot=False, solver=solver_convex)
             self.boolean_masks.daytime = ss.sunup_mask_estimated
-        except:
+        except Exception:
             msg = "Sunrise/sunset detection failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -451,7 +449,7 @@ class DataHandler:
             progress.update()
         try:
             self.make_filled_data_matrix(zero_night=zero_night, interp_day=interp_day)
-        except:
+        except Exception:
             msg = "Matrix filling failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -505,7 +503,7 @@ class DataHandler:
         try:
             # density scoring
             self.get_daily_scores(threshold=0.2, solver=solver_convex)
-        except:
+        except Exception:
             msg = "Daily quality scoring failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -518,7 +516,7 @@ class DataHandler:
                 density_upper_threshold=density_upper_threshold,
                 linearity_threshold=linearity_threshold,
             )
-        except:
+        except Exception:
             msg = "Daily quality flagging failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -532,7 +530,7 @@ class DataHandler:
                 energy_threshold=clear_day_energy_param,
                 solver=solver_convex,
             )
-        except:
+        except Exception:
             msg = "Clear day detection failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -551,7 +549,7 @@ class DataHandler:
         t_clean[3] = time()
         try:
             self.score_data_set()
-        except:
+        except Exception:
             msg = "Data set summary scoring failed."
             self._error_msg += "\n" + msg
             if verbose:
@@ -589,7 +587,7 @@ class DataHandler:
                     periodic_detector=periodic_detector,
                     solver=solver,
                 )
-                rms = lambda x: np.sqrt(np.mean(np.square(x)))
[…]
