+++ statistical_clear_sky/__init__.py
@@ -1,0 +1,13 @@
+from statistical_clear_sky.algorithm.iterative_fitting import IterativeFitting as SCSF
+
+from warnings import warn
+
+warn(
+    "The Statistical Clear Sky package is deprecated. Starting in Solar Data Tools 2.0, it will be removed.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+__all__ = [
+    "SCSF",
+]

[…]
