+++ solardatatools/data_handler.py
@@ -5,6 +5,8 @@ This module contains a class for managing a data processing pipeline
 
 """
 
+from solardatatools.algorithms.dilation import Dilation
+from solardatatools.polar_transform import PolarTransform
 from time import time
 from datetime import timedelta
 from datetime import datetime
@@ -14,7 +16,8 @@ import cvxpy as cvx
 from sklearn.cluster import DBSCAN
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
-import traceback, sys
+import traceback
+import sys
 from tqdm import tqdm
 from solardatatools.time_axis_manipulation import (
     make_time_series,
@@ -43,7 +46,6 @@ from solardatatools.algorithms import (
 from pandas.plotting import register_matplotlib_converters
 
 register_matplotlib_converters()
-from solardatatools.polar_transform import PolarTransform
 
 
 class DataHandler:
@@ -139,7 +141,8 @@ class DataHandler:
         if self.raw_data_matrix is not None:
             self.num_days = self.raw_data_matrix.shape[1]
             if self.raw_data_matrix.shape[0] <= 1400:
-                self.data_sampling = int(24 * 60 / self.raw_data_matrix.shape[0])
+                self.data_sampling = int(
+                    24 * 60 / self.raw_data_matrix.shape[0])
             else:
                 self.data_sampling = 24 * 60 / self.raw_data_matrix.shape[0]
         else:
@@ -426,7 +429,8 @@ class DataHandler:
             progress.update()
         ss = SunriseSunset()
         try:
-            ss.run_optimizer(self.raw_data_matrix, plot=False, solver=solver_convex)
+            ss.run_optimizer(self.raw_data_matrix,
+                             plot=False, solver=solver_convex)
             self.boolean_masks.daytime = ss.sunup_mask_estimated
         except:
             msg = "Sunrise/sunset detection failed."
@@ -442,7 +446,8 @@ class DataHandler:
         if verbose:
             progress.update()
         try:
-            self.make_filled_data_matrix(zero_night=zero_night, interp_day=interp_day)
+            self.make_filled_data_matrix(
+                zero_night=zero_night, interp_day=interp_day)
         except:
             msg = "Matrix filling failed."
             self._error_msg += "\n" + msg
@@ -479,7 +484,8 @@ class DataHandler:
             self._error_msg += "\n" + msg
             if verbose:
                 print(msg)
-                print(f"ratio of filled to raw nonzero measurements was {ratio:.2f}")
+                print(
+                    f"ratio of filled to raw nonzero measurements was {ratio:.2f}")
             self.daily_scores = None
             self.daily_flags = None
             self.data_quality_score = None
@@ -581,7 +587,7 @@ class DataHandler:
                     periodic_detector=periodic_detector,
                     solver=solver,
                 )
-                rms = lambda x: np.sqrt(np.mean(np.square(x)))
+                def rms(x): return np.sqrt(np.mean(np.square(x)))
                 if rms(self.time_shift_analysis.s2) > 0.25:
                     if verbose:
                         print("Invoking periodic timeshift detector.")
@@ -618,7 +624,8 @@ class DataHandler:
                 # Related to this bug fix:
                 # https://github.com/slacgismo/solar-data-tools/commit/ae0037771c09ace08bff5a4904475da606e934da
                 old_index = self.data_frame.index.copy()
-                self.data_frame.index = self.data_frame.index.shift(tz_offset, freq="H")
+                self.data_frame.index = self.data_frame.index.shift(
+                    tz_offset, freq="H")
                 self.data_frame = self.data_frame.reindex(
                     index=old_index, method="nearest", limit=1
                 ).fillna(0)
@@ -627,7 +634,8 @@ class DataHandler:
[…]
