### 2025-06-04 9e1df8f
bundt plotting function in algortihms
