### 2025-06-04 34cb7eb
added aggregate and skip functionalities to bundt cake plot
