### 2025-06-10 #210 @cdeline
Looks good to me.  I pulled changes and tested notebooks for functionality, everything appears to be working.  Thanks for your efforts!
