### 2025-06-10 #210 @jinningwang
Thanks for the improvement and most changes are good to me.

One last concern is: **Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)**

1. With `flake8 7.2.0`, I ran tests on https://github.com/slacgismo/solar-data-tools/commit/792df3f325b87f11690e60848d1915422bf3fe12, some format issues occurred.
If there is a lint config file or linter specified, feel free to let me know

```bash
./sdt_dask/clients/local_client.py:61:80: E501 line too long (80 > 79 characters)
./sdt_dask/clients/local_client.py:62:80: E501 line too long (80 > 79 characters)
./sdt_dask/clients/local_client.py:70:80: E501 line too long (80 > 79 characters)
./sdt_dask/clients/local_client.py:80:80: E501 line too long (93 > 79 characters)
./sdt_dask/clients/local_client.py:85:80: E501 line too long (104 > 79 characters)
./sdt_dask/dask_tool/runner.py:3:80: E501 line too long (82 > 79 characters)
./sdt_dask/dask_tool/runner.py:4:80: E501 line too long (86 > 79 characters)
./sdt_dask/dask_tool/runner.py:75:80: E501 line too long (80 > 79 characters)
./sdt_dask/dask_tool/runner.py:89:80: E501 line too long (82 > 79 characters)
./sdt_dask/dask_tool/runner.py:118:80: E501 line too long (128 > 79 characters)
./sdt_dask/dask_tool/runner.py:135:80: E501 line too long (82 > 79 characters)
./sdt_dask/dask_tool/runner.py:156:80: E501 line too long (82 > 79 characters)
./sdt_dask/dask_tool/runner.py:201:80: E501 line too long (81 > 79 characters)
./sdt_dask/dask_tool/runner.py:214:80: E501 line too long (81 > 79 characters)
./sdt_dask/dataplugs/S3Bucket_plug.py:57:80: E501 line too long (80 > 79 characters)
./sdt_dask/dataplugs/S3Bucket_plug.py:58:80: E501 line too long (80 > 79 characters)
./sdt_dask/dataplugs/S3Bucket_plug.py:75:80: E501 line too long (100 > 79 characters)
./sdt_dask/dataplugs/csv_plug.py:14:80: E501 line too long (81 > 79 characters)
./sdt_dask/dataplugs/csv_plug.py:48:80: E501 line too long (82 > 79 characters)
./sdt_dask/dataplugs/dataplug.py:12:80: E501 line too long (81 > 79 characters)
./sdt_dask/dataplugs/pvdb_plug.py:22:80: E501 line too long (81 > 79 characters)
...
```

2. I didn't find lint process in CI here, https://github.com/slacgismo/solar-data-tools/blob/792df3f325b87f11690e60848d1915422bf3fe12/.github/workflows/lint.yml#L4

A GitHub action might be useful for lint purpose and its integration in CI, https://github.com/marketplace/actions/lint-action
