### 2025-06-10 c950ede
clear sky detection object
