### 2025-06-10 260f432
added clear sky function to data handler
