### 2025-06-11 #210 @pluflou
@cdeline @jinningwang Thank you so much for reviewing the changes!

@jinningwang I am using ruff for formatting, which may have different settings than flake8 (we can control them and make them match, but I left them mostly at default). I set up linting up using pre-commit hooks (you can see the instructions [here on step 7](https://github.com/slacgismo/solar-data-tools/blob/main/CONTRIBUTING.md)), but you can also install ruff and run `ruff check` and `ruff format`. The settings for the ruff checks are in [here](https://github.com/slacgismo/solar-data-tools/blob/792df3f325b87f11690e60848d1915422bf3fe12/.pre-commit-config.yaml#L20). The workflow is running the linting check on this [line](https://github.com/slacgismo/solar-data-tools/blob/792df3f325b87f11690e60848d1915422bf3fe12/.github/workflows/lint.yml#L38). What you are seeing is due to the fact that ruff does not enforce E501 by default, and I did not add it, but it would be straightforward to do so if that's your recommendation. Let me known if you have other questions/your thoughts.
