### 2025-06-11 1b602db
clear sky better reorganized
