### 2025-06-11 f95e942
more fixes to clear sky
