### 2025-06-11 beec17f
Giray final edits to clearsky detection
