### 2025-06-11 03533fe
Update clear_sky_detection.py
