### 2025-06-11 65ba484
modify boolean masks attribute
