+++ solardatatools/data_handler.py
@@ -206,8 +206,6 @@ class DataHandler:
         self.sig_dilated = None
         # Quantile attribute (dictionary)
         self.quantiles = {}
-        # Clear sky detection attributes
-        self.clearsky_masks = None
         # Private attributes
         self._ran_pipeline = False
         self._error_msg = ""
@@ -2489,7 +2487,7 @@ time zone errors     {report['time zone correction'] != 0}
         self.clearsky_sig = csd.get_clearsky_sig()
         daytime = self.boolean_masks.daytime.ravel(order='F')
         self.clearsky_sig[~daytime] = 0
-        self.clearsky_masks = self.clearsky_sig.reshape(
+        self.boolean_masks.clear_times = self.clearsky_sig.reshape(
             self.boolean_masks.daytime.shape, order='F')
 
     def plot_bundt(self,
