### 2025-06-12 #210 @pluflou
Thank you @jinningwang for your review! We really appreciate it.

@cmarmo Could you let me know what the next steps are? We also have a JOSS paper in the repo, but I'm not sure what the process is after this.
