### 2025-06-13 #210 @cmarmo
@pflou, as soon as @echedey-ls will finalise his answer, we can declare the package accepted and move forward with the JOSS submission.
Luckily we have a nice checklist for that too ... :)
