### 2025-06-13 #210 @cmarmo
Wonderful! I guess it's time to move to the next step then.

🎉 Solar Data Tools has been approved by pyOpenSci! Thank you @pluflou  for submitting and many thanks to @cdeline, @echedey-ls and @jinningwang  for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of Solar Data Tools. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [x] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

---
