### 2025-06-13 #210 @cmarmo
@pluflou, @bmeyers, @cdeline, @echedey-ls and @jinningwang, please let me know if you are interested in an invite to the pyopensci Slack, I will be happy to provide you with it.

Also if the authors are interested in writing a blogpost about the review process pyopensci can provide the infrastructure to host it.
