### 2025-06-13 #210 @cmarmo
> I can join it. Do I need to register it first?

@jinningwang you are right, you are already in! No need for an invite! 🙂
