### 2025-06-13 54f13b0
add pyos badge to readme
