### 2025-06-13 #210 @pluflou
Thank you @cmarmo! Please add me to the Slack channel! I update the remaining tasks below, I will update as things progress with JOSS. Who submits the JOSS issue?

* [X]  Add the badge for pyOpenSci peer-review to the README.md of Solar Data Tools. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

* [X]   Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.
