### 2025-06-14 #210 @cmarmo
Thank you @pluflou ! Do you mind creating a new release containing the modifications originated by the review? The 1.7.1 is dated back to April. The new one will be the version accepted by pyOpenSci.

> Who submits the JOSS issue?

The authors / maintainers  of the package take care of the submission: be sure to mention and link to the approved pyOpenSci review. Thanks a lot!
I will take care of the labelling of this issue while the JOSS submission moves forward.
