### 2025-06-19 67ae56d
bug fixes in new code
