+++ solardatatools/algorithms/__init__.py
@@ -9,3 +9,4 @@ from solardatatools.algorithms.soiling import soiling_seperation
 from solardatatools.algorithms.dilation import Dilation
 from solardatatools.algorithms.loss_factor_analysis import LossFactorAnalysis
 from solardatatools.algorithms.quantile_estimation import PVQuantiles
+from solardatatools.algorithms.clear_sky_detection import ClearSkyDetection
