+++ solardatatools/algorithms/clear_sky_detection.py
@@ -11,12 +11,11 @@ import numpy as np
 
 
 class ClearSkyDetection:
-    def __init__(self, data_handler, threshold=0.7, stickiness=2, Q98=None, sig=None,  **config):
-        self.dh = data_handler
+    def __init__(self, sig, quantile_estimate, threshold=0.7, stickiness=2):
         self.stickiness = stickiness
         self.sig = sig
         self.T = self.sig.shape[0]
-        self.Q98 = Q98
+        self.Q98 = quantile_estimate
         self.threshold = threshold
         self.clearsky_sig = np.zeros_like(self.sig, dtype=int)
         self.run()
