### 2025-06-20 2a294e0
fixed and improved dynamic programming algorithm for clear sky state estimation
