### 2025-06-23 b2b6c98
updating DataHandler clear sky functions to use quantile methods rather than SCSF methods
