+++ solardatatools/data_handler.py
@@ -32,7 +32,6 @@ from solardatatools.data_quality import (
 from solardatatools.data_filling import zero_nighttime, interp_missing
 from solardatatools.clear_day_detection import ClearDayDetection
 from solardatatools.plotting import plot_2d
-from solardatatools.clear_time_labeling import find_clear_times
 from solardatatools.solar_noon import avg_sunrise_sunset
 from solardatatools.algorithms.dilation import undilate_signal
 from solardatatools.polar_transform import PolarTransform
@@ -973,82 +972,55 @@ time zone errors     {report['time zone correction'] != 0}
 
     def fit_statistical_clear_sky_model(
         self,
-        data_matrix=None,
-        rank=6,
-        mu_l=None,
-        mu_r=None,
-        tau=None,
-        exit_criterion_epsilon=1e-3,
-        solver_type="MOSEK",
-        max_iteration=10,
-        calculate_degradation=True,
-        max_degradation=None,
-        min_degradation=None,
-        non_neg_constraints=False,
-        verbose=True,
-        bootstraps=None,
+        quantile_level=0.9,
+        nvals_dil=101,
+        num_harmonics=[16, 3], 
+        regularization=0.1, 
+        solver='CLARABEL', 
+        verbose=False
     ):
         """
-        Fit Statistical Clear Sky model.
+        Fit statistical model of PV system clear sky response, using smooth, periodic quantile estimation. Uses new 
+        self.estimate_quantiles method, estimating a single default quantile level of 0.9.
 
-        .. deprecated:: 1.5.0
-            Statistical Clear Sky is deprecated. Starting in Solar Data Tools 2.0, it will be removed.
-
-        :param data_matrix:
-        :param rank:
-        :param mu_l:
-        :param mu_r:
-        :param tau:
-        :param exit_criterion_epsilon:
-        :param solver_type:
-        :param max_iteration:
-        :param calculate_degradation:
-        :param max_degradation:
-        :param min_degradation:
-        :param non_neg_constraints:
-        :param verbose:
-        :param bootstraps:
+        :param quantile_level: the quantile level to fit as the clear sky system response
+        :type quantile_level: float
+        :param nvals_dil: the number of data points to use for daily dilation
+        :type nvals_dil: int
+        :param num_harmonics: the number of Fourier harmonics to use for the daily (first) and yearly (second) periods
+        :type num_harmonics: list
+        :param regularization: stiffness weight for quantile fits (larger is more stiff)
+        :type regularization: float
+        :param solver: the cvxpy solver to invoke, default is CLARABEL
+        :type solver: string
+        :param verbose: print solver status
+        :type verbose: boolean
         :return:
 
         """
-        try:
-            from statistical_clear_sky import SCSF
-        except ImportError:
-            print("Please install statistical-clear-sky package")
-            return
-        scsf = SCSF(
-            data_handler_obj=self,
-            data_matrix=data_matrix,
-            rank_k=rank,
-            solver_type=solver_type,
+        ql = np.atleast_1d(quantile_level)
+        self.estimate_quantiles(
+            nvals_dil=nvals_dil,
+            quantile_levels=ql,
+            num_harmonics=num_harmonics,
+            regularization=regularization,
+            solver=solver,
+            verbose=verbose
         )
-        scsf.execute(
-            mu_l=mu_l,
-            mu_r=mu_r,
-            tau=tau,
-            exit_criterion_epsilon=exit_criterion_epsilon,
-            max_iteration=max_iteration,
-            is_degradation_calculated=calculate_degradation,
-            max_degradation=max_degradation,
-            min_degradation=min_degradation,
[…]
