+++ tests/solardatatools/test_bundt_plot.py
@@ -1,0 +1,13 @@
+import unittest
+import matplotlib.pyplot as plt
+import numpy as np
+from solardatatools.plotting import plot_bundt_cake
+
+class TestBundtPlot(unittest.TestCase):
+    def test_bundt_plot(self):
+        np.random.seed(42)
+        test_data = np.random.rand(365, 100)
+        fig = plot_bundt_cake(test_data)
+
+if __name__ == "__main__":
+    unittest.main()
