+++ solardatatools/data_handler.py
@@ -42,7 +42,7 @@ from solardatatools.algorithms import (
     LossFactorAnalysis,
     PVQuantiles,
     Dilation,
-    ClearSkyDetection
+    ClearSkyDetection,
 )
 from pandas.plotting import register_matplotlib_converters
 
@@ -147,8 +147,7 @@ class DataHandler:
         if self.raw_data_matrix is not None:
             self.num_days = self.raw_data_matrix.shape[1]
             if self.raw_data_matrix.shape[0] <= 1400:
-                self.data_sampling = int(
-                    24 * 60 / self.raw_data_matrix.shape[0])
+                self.data_sampling = int(24 * 60 / self.raw_data_matrix.shape[0])
             else:
                 self.data_sampling = 24 * 60 / self.raw_data_matrix.shape[0]
         else:
@@ -439,8 +438,7 @@ class DataHandler:
             progress.update()
         ss = SunriseSunset()
         try:
-            ss.run_optimizer(self.raw_data_matrix,
-                             plot=False, solver=solver_convex)
+            ss.run_optimizer(self.raw_data_matrix, plot=False, solver=solver_convex)
             self.boolean_masks.daytime = ss.sunup_mask_estimated
         except Exception:
             msg = "Sunrise/sunset detection failed."
@@ -493,8 +491,7 @@ class DataHandler:
             self._error_msg += "\n" + msg
             if verbose:
                 print(msg)
-                print(
-                    f"ratio of filled to raw nonzero measurements was {ratio:.2f}")
+                print(f"ratio of filled to raw nonzero measurements was {ratio:.2f}")
             self.daily_scores = None
             self.daily_flags = None
             self.data_quality_score = None
@@ -633,8 +630,7 @@ class DataHandler:
                 # Related to this bug fix:
                 # https://github.com/slacgismo/solar-data-tools/commit/ae0037771c09ace08bff5a4904475da606e934da
                 old_index = self.data_frame.index.copy()
-                self.data_frame.index = self.data_frame.index.shift(
-                    tz_offset, freq="H")
+                self.data_frame.index = self.data_frame.index.shift(tz_offset, freq="H")
                 self.data_frame = self.data_frame.reindex(
                     index=old_index, method="nearest", limit=1
                 ).fillna(0)
@@ -643,8 +639,7 @@ class DataHandler:
                 self.filled_data_matrix = np.nan_to_num(
                     np.roll(self.filled_data_matrix, roll_by, axis=0), 0
                 )
-                self.raw_data_matrix = np.roll(
-                    self.raw_data_matrix, roll_by, axis=0)
+                self.raw_data_matrix = np.roll(self.raw_data_matrix, roll_by, axis=0)
                 self.boolean_masks.daytime = np.roll(
                     self.boolean_masks.daytime, roll_by, axis=0
                 )
@@ -794,8 +789,7 @@ class DataHandler:
                         self.num_days / 365
                     )
                 else:
-                    l1 = "Length:                {} days\n".format(
-                        self.num_days)
+                    l1 = "Length:                {} days\n".format(self.num_days)
                 if self.power_units == "W":
                     l1_a = "Capacity estimate:     {:.2f} kW\n".format(
                         self.capacity_estimate / 1000
@@ -810,8 +804,7 @@ class DataHandler:
                     )
                     l1_a += self.power_units + "\n"
                 if self.raw_data_matrix.shape[0] <= 1440:
-                    l2 = "Data sampling:         {} minute\n".format(
-                        self.data_sampling)
+                    l2 = "Data sampling:         {} minute\n".format(self.data_sampling)
                 else:
                     l2 = "Data sampling:         {} second\n".format(
[…]
