### 2025-06-25 5a8c755
ruff linting
