+++ solardatatools/algorithms/shade.py
@@ -66,13 +66,25 @@ class ShadeAnalysis:
         return state
 
     def run(
-        self, power=8, solver="MOSEK", w1=20, w2=1e0 , w3=5e2, w4=3e-1, verbose=False, mu=None, lambd=None, q_mat=None
+        self,
+        power=8,
+        solver="MOSEK",
+        w1=20,
+        w2=1e0,
+        w3=5e2,
+        w4=3e-1,
+        verbose=False,
+        mu=None,
+        lambd=None,
+        q_mat=None,
     ):
         if not self.has_run:
             dn, dt = self.transform_data(power)
             self.data_normalized = dn
             self.data_transformed = dt
-            self.osd_problem = self.make_osd_problem(w1=w1, w2=w2, w3=w3, w4=w4, mu=mu, lambd=lambd, q_mat=q_mat)
+            self.osd_problem = self.make_osd_problem(
+                w1=w1, w2=w2, w3=w3, w4=w4, mu=mu, lambd=lambd, q_mat=q_mat
+            )
             self.osd_problem.solve(solver=solver, verbose=verbose)
         variable_dict = {v.name(): v for v in self.osd_problem.variables()}
         self.clear_sky_component = variable_dict["clear-sky"].value / self.scale_factor
@@ -336,7 +348,9 @@ class ShadeAnalysis:
         agg_by_azimuth = agg_by_azimuth.reindex(new_index)
         return normalized, agg_by_azimuth
 
-    def make_osd_problem(self, w1=20, w2=1e0 , w3=5e2, w4=3e-1, mu=None, lambd=None, q_mat=None):
+    def make_osd_problem(
+        self, w1=20, w2=1e0, w3=5e2, w4=3e-1, mu=None, lambd=None, q_mat=None
+    ):
         if mu is None:
             mu = MU
         if lambd is None:
