+++ solardatatools/algorithms/clear_sky_detection.py
@@ -17,14 +17,42 @@ class ClearSkyDetection:
         quantile_estimate,
         threshold_low=0.75,
         threshold_high=1.2,
-        stickiness_high=0.1,
         stickiness_low=4,
+        stickiness_high=0.1,
     ):
+        """
+        A class for detecting clear sky periods in measured PV power data. This class requires the user
+        to provide a signal to be labeled, a time-series quantile estimate of that signal, and a number
+        of parameters that control the dynamic programming algorithm.
+
+        The basic model is that the ratio of the measured signal to the quantile estimate is between a band
+        defined by threshold_low and threshold_high when conditions are "clear sky", and the ratio
+        is outside that band when conditions are not "clear sky". The "stickiness" parameters define the cost
+        to transition out of a cloudy state (stickiness_low) or out of a clear state (stickiness_high). We
+        set stickiness_low > stickiness_high so that it is easier to transition from clear to cloudy than from
+        cloudy to clear.
+
+        :param sig: the measured PV power signal to be labeled
+        :type sig: 1D numpy.array object
+        :param quantile_estimate: the estimated "clear sky" quantile, typically the 0.9 level
+        :type sig: 1D numpy.array object of same length as sig
+        :param theshold_low: the lower boundry of the clear sky ratio band
+        :type theshold_low: float
+        :param theshold_high: the upper boundry of the clear sky ratio band
+        :type theshold_high: float
+        :param stickiness_low: the cost to transition out of the cloudy state (large)
+        :type stickiness_low: float
+        :param stickiness_high: the cost to transition out of the clear sky state (small)
+        :type stickiness_high: float
+        """
+        assert len(sig) == len(quantile_estimate), (
+            "signal and quantile estimate must have the same length!"
+        )
         self.stickiness_low = stickiness_low
         self.stickiness_high = stickiness_high
-        self.sig = sig
+        self.sig = np.atleast_1d(sig)
         self.T = self.sig.shape[0]
-        self.Q98 = quantile_estimate
+        self.Q98 = np.atleast_1d(quantile_estimate)
         self.threshold_low = threshold_low
         self.threshold_high = threshold_high
         self.clearsky_sig = np.zeros_like(self.sig, dtype=int)
