### 2025-06-26 8ead49b
default plot behavior should be to aggregate
