### 2025-06-26 c7c23c8
fix whitespace in bundt plots
