### 2025-06-26 ee01b46
kwargs were not being passed to inner scope correctly
