+++ docs/source/reference/solardatatools.algorithms.rst
@@ -16,6 +16,8 @@ Submodules
    solardatatools.algorithms.clipping
    solardatatools.algorithms.dilation
    solardatatools.algorithms.loss_factor_analysis
+   solardatatools.algorithms.clear_sky_detection
+   solardatatools.algorithms.quantile_estimation
    solardatatools.algorithms.shade
    solardatatools.algorithms.soiling
    solardatatools.algorithms.sunrise_sunset_estimation
