### 2025-06-26 9a2dfe1
Deleting Statistical Clear Sky Fitting algorithm based on bi-convex minimization and matrix factorization. New approach based on smooth, multi-periodic quantile estimation has already been implemented.
