+++ docs/source/reference/solardatatools.algorithms.clear_sky_detection.rst
@@ -1,5 +1,5 @@
 Clear Sky Detection
-===========
+===================
 
 .. automodule:: solardatatools.algorithms.clear_sky_detection
    :members:
