+++ docs/source/reference/solardatatools.algorithms.quantile_estimation.rst
@@ -1,5 +1,5 @@
 Time-Series (Smooth, Multi-Periodic) Quantile Estimation
-===========
+========================================================
 
 .. automodule:: solardatatools.algorithms.quantile_estimation
    :members:
