### 2025-06-27 cc32050
fix affiliation and DOIs
