+++ paper/paper.bib
@@ -43,19 +43,19 @@ month = {9}
 	month        = August,
 	year         = 2024,
 	version      = {v3.0.0},
-	url          = {https://doi.org/10.5281/zenodo.13356337}
+	url          = {10.5281/zenodo.13356337}
 }
 
 
 @misc{sdt-zenodo,
   author       = {Meyers, Bennet and others},
-  title        = {slacgismo/solar-data-tools: v1.2.2},
+  title        = {slacgismo/solar-data-tools: v1.7.1},
   month        = mar,
   year         = 2024,
   publisher    = {Zenodo},
-  version      = {v1.2.2},
-  doi          = {10.5281/zenodo.10888385},
-  url          = {https://doi.org/10.5281/zenodo.10888385}
+  version      = {v1.7.1},
+  doi          = {10.5281/zenodo.15243716},
+  url          = {10.5281/zenodo.15243716}
 }
 
 @misc{sdt-repo,
@@ -63,7 +63,7 @@ month = {9}
 	title        = {Solar Data Tools},
 	month        = mar,
 	year         = 2024,
-	version      = {v1.2.2},
+	version      = {v1.7.1},
 	url          = {https://github.com/slacgismo/solar-data-tools}
 }
 
@@ -145,7 +145,7 @@ volume = {29},
 number = {6},
 pages = {573-602},
 keywords = {degradation rate, performance loss rate, PV system degradation, PV system performance},
-doi = {https://doi.org/10.1002/pip.3397},
+doi = {10.1002/pip.3397},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pip.3397},
 year = {2021}
 }
@@ -167,7 +167,7 @@ volume = {236},
 pages = {121493},
 year = {2024},
 issn = {0960-1481},
-doi = {https://doi.org/10.1016/j.renene.2024.121493},
+doi = {10.1016/j.renene.2024.121493},
 url = {https://www.sciencedirect.com/science/article/pii/S0960148124015611},
 author = {Baojie Li and Xin Chen and Anubhav Jain},
 }
@@ -179,7 +179,7 @@ volume = {254},
 pages = {168-181},
 year = {2023},
 issn = {0038-092X},
-doi = {https://doi.org/10.1016/j.solener.2023.03.011},
+doi = {10.1016/j.solener.2023.03.011},
 url = {https://www.sciencedirect.com/science/article/pii/S0038092X23001573},
 author = {Baojie Li and Todd Karin and Bennet E. Meyers and Xin Chen and Dirk C. Jordan and Clifford W. Hansen and Bruce H. King and Michael G. Deceglie and Anubhav Jain},
 }
