### 2025-06-27 #210 @pluflou
The [paper](https://joss.theoj.org/papers/10.21105/joss.08478) has been [approved by JOSS](https://github.com/openjournals/joss-reviews/issues/8478). Thank you everyone! 🎉
